% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apg.R
\name{apg}
\alias{apg}
\alias{get_apg_orders}
\alias{get_apg_families}
\title{Get APG names}
\usage{
get_apg_orders(...)

get_apg_families(...)
}
\arguments{
\item{...}{Curl args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Generic names and their replacements from the Angiosperm Phylogeny
Group III system of flowering plant classification.
}
\examples{
\dontrun{
get_apg_orders()
get_apg_families()
}
}
\references{
http://www.mobot.org/MOBOT/research/APweb/
}
