% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{apg_families}
\alias{apg_families}
\title{MOBOT family names}
\format{A data frame with 1597 rows and 4 variables:
\itemize{
\item \code{original} original data record from APG website
\item \code{this} Order name
\item \code{that} Replacement order name
\item \code{order} Order name
}}
\source{
\url{http://www.mobot.org/MOBOT/research/APweb/}
}
\description{
Family names and their replacements from the Angiosperm Phylogeny
Website system of flowering plant classification.
}
\details{
This dataset is from Version 13, incorporated on 2015-04-29.
}
\keyword{data}
