% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{apg_genera}
\alias{apg_genera}
\title{APGIII generic names}
\format{A data frame with 20,075 rows and 5 variables:
\describe{
  \item{family}{Family name}
  \item{name}{Genus name}
  \item{replace}{Replacment genus name}
  \item{name_authority}{Authority}
  \item{replace_authority}{Replacement authority}
}}
\source{
\url{http://www.mobot.org/MOBOT/research/APweb/}
}
\description{
Generic names and their replacements from the Angiosperm Phylogeny
Group III system of flowering plant classification.
}
\details{
This dataset is from Version 13, incorporated on 2015-04-29.
}
\keyword{data}

