% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apni_utils.R
\name{apni}
\alias{apni}
\alias{apni_search}
\alias{apni_classification}
\alias{apni_children}
\alias{apni_family}
\alias{acceptable_names}
\title{Search Australian Plant Names Index}
\usage{
apni_search(q, ...)

apni_classification(id, ...)

apni_children(id, ...)

apni_family(id, ...)

acceptable_names(q, ...)
}
\arguments{
\item{q}{(character) name to query}

\item{...}{Further args passed on to \link[crul:HttpClient]{crul::HttpClient}.}
}
\value{
a list with slots for metadata (\code{meta}) with list of response
attributes, and data (\code{data}) with a data.frame of results
}
\description{
Search Australian Plant Names Index
}
\examples{
\dontrun{
x <- apni_search(q = "Pinus")
x
x$accepted_names
x$synonyms
x$synonyms$acceptedNameUsage
x$meta$perPage

apni_search(q = "Acacia")

# Acacia dealbata 
apni_classification(id = 61294)
# Acacia 
apni_classification(id = 56859)

# children
apni_children(id = 51311124)

# family?
apni_family(id = 158548)

# acceptable names
acceptable_names(q = 'Poa fax')
acceptable_names(q = 'Poa fa')
acceptable_names(q = 'Poa')
}
}
\references{
https://biodiversity.org.au/nsl/docs/main.html
}
\seealso{
Other apni: 
\code{\link{get_apni}()}
}
\concept{apni}
