% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apni_utils.R
\name{apni_children}
\alias{apni_children}
\title{Australian Plant Names Index children}
\usage{
apni_children(id, ...)
}
\arguments{
\item{id}{(numeric/character) one or more APNI ids}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}.}
}
\value{
a list with slots for name (\code{name}), link (\code{link}),
and the taxonomic children (\code{children})
}
\description{
Australian Plant Names Index children
}
\examples{
\dontrun{
x <- apni_children(id = 56859)
x[[1]]
x[[1]]$name
x[[1]]$link
x[[1]]$children
x[[1]]$children$id
}
}
\references{
https://biodiversity.org.au/nsl/docs/main.html
}
\seealso{
Other apni: 
\code{\link{apni-search}},
\code{\link{apni_classification}()},
\code{\link{apni_family}()},
\code{\link{apni_id}()},
\code{\link{get_apni}()}
}
\concept{apni}
