% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apni_utils.R
\name{apni_classification}
\alias{apni_classification}
\title{Australian Plant Names Index classification}
\usage{
apni_classification(id, ...)
}
\arguments{
\item{id}{(numeric/character) one or more APNI ids}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}.}
}
\value{
a list with slots for name (\code{name}), link (\code{link}),
and the taxonomic classification (\code{hierarchy})
}
\description{
Australian Plant Names Index classification
}
\examples{
\dontrun{
# Acacia dealbata 
apni_classification(id = 61294)
# Acacia 
apni_classification(id = 56859)
}
}
\references{
https://biodiversity.org.au/nsl/docs/main.html
}
\seealso{
Other apni: 
\code{\link{apni-search}},
\code{\link{apni_children}()},
\code{\link{apni_family}()},
\code{\link{apni_id}()},
\code{\link{get_apni}()}
}
\concept{apni}
