% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apni_utils.R
\name{apni_family}
\alias{apni_family}
\title{Australian Plant Names Index family information}
\usage{
apni_family(id, ...)
}
\arguments{
\item{id}{(numeric/character) one or more APNI ids}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}.}
}
\value{
a list with slots for name (\code{name}), link (\code{link}),
and the instances (\code{instances})
}
\description{
Returns the family of the name according to the APNI or
'Name classification'. The Name classification may be different
to other classifications such as APC.
}
\examples{
\dontrun{
apni_family(id = 158548)
}
}
\references{
https://biodiversity.org.au/nsl/docs/main.html
}
\seealso{
Other apni: 
\code{\link{apni-search}},
\code{\link{apni_children}()},
\code{\link{apni_classification}()},
\code{\link{apni_id}()},
\code{\link{get_apni}()}
}
\concept{apni}
