% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apni_utils.R
\name{apni_id}
\alias{apni_id}
\alias{apni_id_exists}
\title{Australian Plant Names Index ID lookup}
\usage{
apni_id(id, ...)

apni_id_exists(id)
}
\arguments{
\item{id}{(numeric/character) one or more APNI ids}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}.}
}
\value{
for \code{apni_id()} a list with lots of data. for
\code{apni_id_exists()} a boolean (\code{FALSE} if the id does not exist).
\code{apni_id_exists()} uses \code{apni_id()} internally
}
\description{
Get APNI metadata for an APNI ID or check that an ID exists
}
\examples{
\dontrun{
# fetch an id
apni_id(56859)
# apni_id(3333333) # does not exist, fails

# check that an id exists within APNI
apni_id_exists(56859) # exists
apni_id_exists(3333333) # does not exist
apni_id_exists(c(56859, 3333333))
}
}
\references{
https://biodiversity.org.au/nsl/docs/main.html
}
\seealso{
Other apni: 
\code{\link{apni-search}},
\code{\link{apni_children}()},
\code{\link{apni_classification}()},
\code{\link{apni_family}()},
\code{\link{get_apni}()}
}
\concept{apni}
