% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backend.R
\name{backend}
\alias{backend}
\alias{backend_get}
\alias{backend_set}
\title{Set location of data acquisition.}
\usage{
backend_set(itis = "api", col = itis, ncbi = itis, theplantlist = itis,
  path = "~/.taxize_local")

backend_get()
}
\arguments{
\item{itis,col,ncbi,theplantlist}{(character) The backend to query data from. Defaults to
\code{api}, which means we query resources on the web. Alternatively, use \code{localsql},
which uses local versions of databases on your own machine. This option does some checks
to make sure things are setup correctly.}

\item{path}{The path to store local databases. Individual databases are within this base
path.}
}
\description{
Set location of data acquisition.
}
\details{
All other data source parameters in \code{backend_set} inherit from the first
\code{itis}, which is by default set to \code{itis = "api"}

Note that the only option for Theplantlist data is "localsql". Alternatively, you can
download raw csv files of their data using \code{\link[taxize]{tpl_get}}, or use
the \code{Taxonstand} package that downloads csv files and uses regex locally in R.
}
\examples{
\dontrun{
# set all to remote api
backend_set("api")
backend_get()

# set all to localsql
backend_set("localsql")
backend_get()

# set individual data sources to different settings
backend_set(col = "localsql", ncbi = "localsql")
backend_get()
}
}

