% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_search.R
\name{bold_children}
\alias{bold_children}
\title{Barcode of Life taxonomic children}
\usage{
bold_children(id, ...)
}
\arguments{
\item{id}{(integer) A BOLD taxonomic identifier. length=1. required}

\item{...}{named curl options passed on to \link[crul:verb-GET]{crul::verb-GET}
debugging}
}
\value{
list of data.frame's
}
\description{
BEWARE: this function scrapes data from the BOLD website, so may
be unstable. That is, one day it may work, and the next it may fail.
Open an issue if you encounter an error:
https://github.com/ropensci/taxize/issues
}
\examples{
\dontrun{
# Osmia (genus): 253 children
bold_children(id = 4940)
# Momotus (genus): 3 children
bold_children(id = 88899)
# Momotus aequatorialis (species): no children
bold_children(id = 115130)
# Osmia sp1 (species): no children
bold_children(id = 293378)
# Arthropoda (phylum): 27 children
bold_children(id = 82)
# Psocodea (order): 51 children
bold_children(id = 737139)
# Megachilinae (subfamily): 2 groups (tribes: 3, genera: 60)
bold_children(id = 4962)
# Stelis (species): 78 taxa
bold_children(id = 4952)
}
}
\keyword{internal}
