% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bold_search}
\alias{bold_search}
\title{Search Barcode of Life for taxonomic IDs}
\usage{
bold_search(name = NULL, id = NULL, fuzzy = FALSE, dataTypes = "basic",
  includeTree = FALSE, response = FALSE, ...)
}
\arguments{
\item{name}{(character) One or more scientific names.}

\item{id}{(integer) One or more BOLD taxonomic identifiers.}

\item{fuzzy}{(logical) Whether to use fuzzy search or not (default: FALSE).}

\item{dataTypes}{(character) Specifies the datatypes that will be returned. 'all' returns all data.
'basic' returns basic taxon information. 'images' returns specimen images.}

\item{includeTree}{(logical) If TRUE (default: FALSE), returns a list containing information
for parent taxa as well as the specified taxon.}
}
\value{
A list of data.frame's.
}
\description{
Search Barcode of Life for taxonomic IDs
}
\details{
You must provide one of name or id. The other parameters (format
		and start) are optional.
}
\examples{
\dontrun{
# A basic example
bold_search(name="Apis")
bold_search(name="Agapostemon")
bold_search(name="Poa")

# Fuzzy search
head(bold_search(name="Po", fuzzy=TRUE))
head(bold_search(name="Aga", fuzzy=TRUE))

# Many names
bold_search(name=c("Apis","Puma concolor"))
nms <- names_list('species')
bold_search(name=nms)

# An example where there is no data
bold_search(id=88899)
}
}
\references{
\url{http://webservice.catalogueoflife.org/}
}

