% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_copied_code.R
\name{bold_tax_id}
\alias{bold_tax_id}
\title{Search BOLD for taxonomy data by BOLD ID.}
\usage{
bold_tax_id(
  id,
  dataTypes = "basic",
  includeTree = FALSE,
  response = FALSE,
  ...
)
}
\description{
Search BOLD for taxonomy data by BOLD ID.
}
\keyword{internal}
