% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{col-defunct}
\alias{col-defunct}
\alias{as.colid}
\alias{col_children}
\alias{col_classification}
\alias{col_downstream}
\alias{col_search}
\alias{get_colid}
\alias{get_colid_}
\title{COL defunct functions}
\usage{
as.colid(...)

col_children(...)

col_classification(...)

col_downstream(...)

col_search(...)

get_colid(...)

get_colid_(...)
}
\arguments{
\item{...}{ignored}
}
\description{
COL introduced rate limiting recently in 2019 - which has made the API
essentially unusable - COL+ is coming soon and we'll incorporate it here
when it's stable
}
\keyword{internal}
