% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_classification.R
\name{col_classification}
\alias{col_classification}
\title{Search Catalogue of Life for taxonomic classifications.}
\usage{
col_classification(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\keyword{internal}
