% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\alias{db_itis}
\title{Download ITIS data}
\usage{
db_itis(verbose = TRUE)
}
\arguments{
\item{verbose}{(logical) Print messages. Default: TRUE}
}
\value{
Downloads sql database, loads it into SQLite, cleans up by removing unneeded files,
and gives back path to the database
}
\description{
Download ITIS data
}
\examples{
\dontrun{
db_itis()
}
}

