% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol_invasive.R
\name{eol_invasive}
\alias{eol_invasive}
\title{Search for presence of taxonomic names in EOL invasive species databases.}
\usage{
eol_invasive(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\keyword{internal}

