% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eol_ping}
\alias{eol_ping}
\title{Ping EOL API to see if it's working.}
\usage{
eol_ping()
}
\value{
XML object message about API status.
}
\description{
Ping EOL API to see if it's working.
}
\examples{
\donttest{
eol_ping()
}
}
\seealso{
\code{\link{itis_ping}}
}

