% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eubon_capabilities.R
\name{eubon_capabilities}
\alias{eubon_capabilities}
\title{EUBON capabilites}
\usage{
eubon_capabilities(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
EUBON capabilites
}
\examples{
\dontrun{
eubon_capabilities()
}
}
\references{
\url{http://cybertaxonomy.eu/eu-bon/utis/1.2/doc.html}
}
\seealso{
Other eubon.methods: \code{\link{eubon_children}},
  \code{\link{eubon_hierarchy}}, \code{\link{eubon}}
}

