\docType{data}
\name{fresh_desc}
\alias{fresh_desc}
\title{Lookup-table for fresh_traits}
\description{
  Lookup-table for fresh_traits. A table with three
  columns: \describe{ \item{"Modality"}{Equates to the
  column-names of \link{fresh_traits}}
  \item{"Description"}{Description of each Modality}
  \item{"Trait"}{Traits} }
}
\examples{
data(fresh_desc)
head(fresh_desc)
}
\keyword{data}
\keyword{datasets}

