% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genank2uid}
\alias{genank2uid}
\title{Get NCBI taxonomy UID from GenBankID}
\usage{
genank2uid(id, ...)
}
\arguments{
\item{id}{A GenBankID}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get NCBI taxonomy UID from GenBankID
}
\examples{
\donttest{
genank2uid(id = 'AJ748748')
}
}

