% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apni.R
\name{get_apni}
\alias{get_apni}
\alias{as.apni}
\alias{as.apni.apni}
\alias{as.apni.numeric}
\alias{as.apni.character}
\alias{as.apni.list}
\alias{as.apni.data.frame}
\alias{as.data.frame.apni}
\alias{get_apni_}
\title{Get Australian Plant Names Index identifier for a taxon}
\usage{
get_apni(sci, ask = TRUE, messages = TRUE, rows = NA, rank_filter = NULL, ...)

as.apni(x, check = TRUE)

\method{as.apni}{apni}(x, check = TRUE)

\method{as.apni}{numeric}(x, check = TRUE)

\method{as.apni}{character}(x, check = TRUE)

\method{as.apni}{list}(x, check = TRUE)

\method{as.apni}{data.frame}(x, check = TRUE)

\method{as.data.frame}{apni}(x, ...)

get_apni_(x, messages = TRUE, rows = NA, ...)
}
\arguments{
\item{sci}{character; A vector of scientific names. Or, a
\code{taxon_state} object (see \link{taxon-state})}

\item{ask}{logical; should \code{get_apni} be run in interactive mode?
If \code{TRUE} and more than one pow is found for teh species, the user is
asked for input. If \code{FALSE} NA is returned for multiple matches.}

\item{messages}{logical; should progress be printed?}

\item{rows}{numeric; Any number from 1 to infinity. If the default NA,
all rows are considered. Note that this function still only gives back
a pow class object with one to many identifiers. See
\code{\link[=get_apni_]{get_apni_()}} to get back all, or a subset,
of the raw data that you are presented during the ask process.}

\item{rank_filter}{(character) A taxonomic rank name to filter data after
retrieved from APNI. See \code{\link[=rank_ref]{rank_ref()}} for possible options.
Though note that some data sources use atypical ranks, so inspect the data
itself for options. Optional. See \code{Filtering} below.}

\item{...}{Curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}

\item{x}{Input to \code{\link[=as.apni]{as.apni()}}}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.apni]{as.apni()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Get Australian Plant Names Index identifier for a taxon
}
\section{Filtering}{

The parameter \code{rank_filter} is not used in the search to the data
provider, but is used in filtering the data down to a subset that is
closer to the target you want.  For these two parameters,
you can use regex strings since we use \code{\link[=grep]{grep()}} internally to match.
Filtering narrows down to the set that matches your query, and removes
the rest.
}

\section{Pagination}{

APNI does not allow pagination. It appears that 100 is the max number
of results that can be returned. That is, if there's more than 100
results \emph{found}, you can only get 100, and no more. So the only
way around it is to refine/change your query.
}

\examples{
\dontrun{
get_apni("Pinus")
get_apni("Acacia ab")
get_apni(c("Acacia","Pinus"))

# Get back a subset
get_apni("Acacia a", rows = 1)
get_apni("Acacia a", rows = 1:10)

# When not found
get_apni("howdy")
get_apni(c("Acacia acicularis", "howdy"))

# Narrow down results 
# to a rank
get_apni("Acacia a", rank_filter = "species", rows = 1:10)

# Convert a apni without class information to a apni class
z <- get_apni("Acacia abbreviata")
# already a apni, returns the same
as.apni(z)
as.apni(57932)
as.apni("57932")
# character vector, length > 1
ids <- c("57932","57943")
as.apni(ids)
# list, with character strings
as.apni(as.list(ids)) 
## dont check, much faster
as.apni("57932", check=FALSE)
as.apni(ids, check=FALSE)
as.apni(as.list(ids), check=FALSE)

(out <- as.apni(ids))
data.frame(out)
as.apni( data.frame(out) )

# Get all data back
get_apni_("Quercus")
get_apni_("Acacia")
get_apni_("Quercus", rows=1:5)
get_apni_("Quercus", rows=1)
get_apni_(c("Pinus", "Abies"), rows = 1:3)
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other apni: 
\code{\link{apni-search}},
\code{\link{apni_children}()},
\code{\link{apni_classification}()},
\code{\link{apni_family}()},
\code{\link{apni_id}()}

Other taxonomic-ids: 
\code{\link{get_boldid}()},
\code{\link{get_eolid}()},
\code{\link{get_gbifid}()},
\code{\link{get_ids}()},
\code{\link{get_iucn}()},
\code{\link{get_natservid}()},
\code{\link{get_nbnid}()},
\code{\link{get_pow}()},
\code{\link{get_tolid}()},
\code{\link{get_tpsid}()},
\code{\link{get_tsn}()},
\code{\link{get_uid}()},
\code{\link{get_wiki}()},
\code{\link{get_wormsid}()}
}
\concept{apni}
\concept{taxonomic-ids}
