% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boldid.R
\name{get_bold}
\alias{get_bold}
\alias{get_boldid}
\alias{as.bold}
\alias{as.bold.bold}
\alias{as.bold.character}
\alias{as.bold.list}
\alias{as.bold.numeric}
\alias{as.bold.data.frame}
\alias{get_bold_}
\alias{get_boldid_}
\title{Get the BOLD (Barcode of Life) code for a search term.}
\usage{
get_bold(
  sci,
  fuzzy = FALSE,
  dataTypes = "basic",
  includeTree = FALSE,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  rank = NULL,
  division = NULL,
  parent = NULL,
  searchterm = NULL,
  ...
)

get_boldid(...)

as.bold(x, check = TRUE)

\method{as.bold}{bold}(x, check = TRUE)

\method{as.bold}{character}(x, check = TRUE)

\method{as.bold}{list}(x, check = TRUE)

\method{as.bold}{numeric}(x, check = TRUE)

\method{as.bold}{data.frame}(x, check = TRUE)

get_bold_(
  sci,
  messages = TRUE,
  fuzzy = FALSE,
  dataTypes = "basic",
  includeTree = FALSE,
  rows = NA,
  searchterm = NULL,
  ...
)

get_boldid_(...)
}
\arguments{
\item{sci}{character; A vector of scientific names. Or,
a \code{taxon_state} object (see \link{taxon-state})}

\item{fuzzy}{(logical) Whether to use fuzzy search or not (default: FALSE).}

\item{dataTypes}{(character) Specifies the datatypes that will be returned.
See \code{\link[=bold_search]{bold_search()}} for options.}

\item{includeTree}{(logical) If TRUE (default: FALSE), returns a list
containing information for parent taxa as well as the specified taxon.}

\item{ask}{logical; should get_itis be run in interactive mode?
If TRUE and more than one TSN is found for teh species, the user is asked for
input. If FALSE NA is returned for multiple matches.}

\item{messages}{logical; should progress be printed?}

\item{rows}{numeric; Any number from 1 to infinity. If the default NA, all rows are
considered. Note that this function still only gives back a bold class object with one
to many identifiers. See \code{\link[=get_bold_]{get_bold_()}} to get back all, or a subset,
of the raw data that you are presented during the ask process.}

\item{rank}{(character) A taxonomic rank name. See \link{rank_ref} for possible
options. Though note that some data sources use atypical ranks, so inspect the
data itself for options. Optional. See \code{Filtering} below.}

\item{division}{(character) A division (aka phylum) name. Optional. See \code{Filtering}
below.}

\item{parent}{(character) A parent name (i.e., the parent of the target search
taxon). Optional. See \code{Filtering} below.}

\item{searchterm}{Deprecated, see \code{sci}}

\item{...}{Curl options passed on to \code{\link[crul:verb-GET]{crul::verb-GET}}}

\item{x}{Input to \code{\link[=as.bold]{as.bold()}}}

\item{check}{logical; Check if ID matches any existing on the DB, only used in
\code{\link[=as.bold]{as.bold()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Get the BOLD (Barcode of Life) code for a search term.
}
\section{Filtering}{

The parameters \code{division}, \code{parent}, and \code{rank} are not used in the search to
the data provider, but are used in filtering the data down to a subset that
is closer to the target you want.  For all these parameters, you can use
regex strings since we use \code{\link[=grep]{grep()}} internally to match. Filtering narrows
down to the set that matches your query, and removes the rest.
}

\examples{
\dontrun{
get_bold(sci = "Agapostemon")
get_bold(sci = "Chironomus riparius")
get_bold(c("Chironomus riparius","Quercus douglasii"))
splist <- names_list('species')
get_bold(splist, messages=FALSE)

# Fuzzy searching
get_bold(sci="Osmi", fuzzy=TRUE)

# Get back a subset
get_bold(sci="Osmi", fuzzy=TRUE, rows = 1)
get_bold(sci="Osmi", fuzzy=TRUE, rows = 1:10)
get_bold(sci=c("Osmi","Aga"), fuzzy=TRUE, rows = 1)
get_bold(sci=c("Osmi","Aga"), fuzzy=TRUE, rows = 1:3)

# found
get_bold("Cordulegaster erronea")
get_bold('Arigomphus furcifer')

# When not found
get_bold('Epicordulia princeps')
get_bold("howdy")
get_bold(c("Chironomus riparius", "howdy"))
get_bold("Nasiaeshna pentacantha")

# Narrow down results to a division or rank, or both
## Satyrium example
### Results w/o narrowing
get_bold("Satyrium")
### w/ phylum
get_bold("Satyrium", division = "Plantae")
get_bold("Satyrium", division = "Animalia")

## Rank example
get_bold("Osmia", fuzzy = TRUE)
get_bold("Osmia", fuzzy = TRUE, rank = "genus")

# Fuzzy filter on any filtering fields
## uses grep on the inside
get_bold("Satyrium", division = "anim")
get_bold("Aga", fuzzy = TRUE, parent = "*idae")

# Convert a bold without class information to a bold class
as.bold(get_bold("Agapostemon")) # already a bold, returns the same
as.bold(get_bold(c("Agapostemon","Quercus douglasii"))) # same
as.bold(1973) # numeric
as.bold(c(1973,101009,98597)) # numeric vector, length > 1
as.bold("1973") # character
as.bold(c("1973","101009","98597")) # character vector, length > 1
as.bold(list("1973","101009","98597")) # list, either numeric or character
## dont check, much faster
as.bold("1973", check=FALSE)
as.bold(1973, check=FALSE)
as.bold(c("1973","101009","98597"), check=FALSE)
as.bold(list("1973","101009","98597"), check=FALSE)

(out <- as.bold(c(1973,101009,98597)))
data.frame(out)
as.bold( data.frame(out) )

# Get all data back
get_bold_("Osmia", fuzzy=TRUE, rows=1:5)
get_bold_("Osmia", fuzzy=TRUE, rows=1)
get_bold_(c("Osmi","Aga"), fuzzy=TRUE, rows = 1:3)
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other taxonomic-ids: 
\code{\link{get_eol}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_itis}()},
\code{\link{get_iucn}()},
\code{\link{get_natserv}()},
\code{\link{get_nbn}()},
\code{\link{get_ncbi}()},
\code{\link{get_pow}()},
\code{\link{get_tol}()},
\code{\link{get_tps}()},
\code{\link{get_wiki}()},
\code{\link{get_worms}()},
\code{\link{id-accessors}}
}
\concept{taxonomic-ids}
