% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eolid.R
\name{get_eol}
\alias{get_eol}
\alias{get_eolid}
\alias{as.eol}
\alias{as.eol.txid}
\alias{as.eol.character}
\alias{as.eol.list}
\alias{as.eol.numeric}
\alias{as.eol.data.frame}
\alias{get_eol_}
\alias{get_eolid_}
\title{Get the EOL ID from Encyclopedia of Life from taxonomic names.}
\usage{
get_eol(
  sci_com,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  rank = NULL,
  data_source = NULL,
  sciname = NULL,
  ...
)

get_eolid(...)

as.eol(x, check = TRUE)

\method{as.eol}{txid}(x, check = TRUE)

\method{as.eol}{character}(x, check = TRUE)

\method{as.eol}{list}(x, check = TRUE)

\method{as.eol}{numeric}(x, check = TRUE)

\method{as.eol}{data.frame}(x, check = TRUE)

get_eol_(sci_com, messages = TRUE, rows = NA, sciname = NULL, ...)

get_eolid_(...)
}
\arguments{
\item{sci_com}{character; one or more scientific or common names. Or,
a \code{taxon_state} object (see \link{taxon-state})}

\item{ask}{logical; should get_eol be run in interactive mode?
If TRUE and more than one ID is found for the species, the user is asked for
input. If FALSE NA is returned for multiple matches.}

\item{messages}{logical; If \code{TRUE} the actual taxon queried is printed
on the console.}

\item{rows}{numeric; Any number from 1 to infinity. If the default NA, all
rows are considered. Note that this function still only gives back a eol
class object with one to many identifiers. See
\code{\link[=get_eol_]{get_eol_()}} to get back all, or a subset, of the raw
data that you are presented during the ask process.}

\item{rank}{(character) A taxonomic rank name. See \link{rank_ref}
for possible options. Though note that some data sources use atypical ranks,
so inspect the data itself for options. Optional. See \code{Filtering} below.}

\item{data_source}{(character) A data source inside of EOL. These are
longish names like e.g., "Barcode of Life Data Systems" or
"USDA PLANTS images". Optional. See \code{Filtering} below.}

\item{sciname}{Deprecated, see \code{sci_com}}

\item{...}{Further args passed on to \code{\link[=eol_search]{eol_search()}}}

\item{x}{Input to \code{\link[=as.eol]{as.eol()}}}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.eol]{as.eol()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Note that EOL doesn't expose an API endpoint for directly querying for EOL
taxon ID's, so we first use the function \code{\link[=eol_search]{eol_search()}}
to find pages that deal with the species of interest, then use
\code{\link[=eol_pages]{eol_pages()}} to find the actual taxon IDs.
}
\details{
EOL is a bit odd in that they have page IDs for each taxon, but
then within that, they have taxon ids for various taxa within that page
(e.g., GBIF and NCBI each have a taxon they refer to within the
page [i.e., taxon]). And we need the taxon ids from a particular data
provider (e.g, NCBI) to do other things, like get a higher classification
tree. However, humans want the page id, not the taxon id. So, the
id returned from this function is the taxon id, not the page id. You can
get the page id for a taxon by using \code{\link[=eol_search]{eol_search()}} and
`\code{\link[=eol_pages]{eol_pages()}}, and the URI returned in the attributes for a
taxon will lead you to the taxon page, and the ID in the URL is the
page id.
}
\section{Filtering}{

The parameters \code{rank} and \code{data_source} are not used in the
search to the data provider, but are used in filtering the data down to a
subset that is closer to the target you want.  For all these parameters,
you can use regex strings since we use \code{\link[=grep]{grep()}} internally to
match. Filtering narrows down to the set that matches your query, and
removes the rest.
}

\examples{
\dontrun{
get_eol(sci_com='Pinus contorta')
get_eol(sci_com='Puma concolor')

get_eol(c("Puma concolor", "Pinus contorta"))

# specify rows to limit choices available
get_eol('Poa annua')
get_eol('Poa annua', rows=1)
get_eol('Poa annua', rows=2)
get_eol('Poa annua', rows=1:2)

# When not found
get_eol(sci_com="uaudnadndj")
get_eol(c("Chironomus riparius", "uaudnadndj"))

# filter results to a rank or data source, or both
get_eol("Satyrium")
get_eol("Satyrium", rank = "genus")
get_eol("Satyrium", data_source = "inaturalist")
get_eol("Satyrium", rank = "genus",
  data_source = "North Pacific Species List")

# Convert a eol without class information to a eol class
# already a eol, returns the same
as.eol(get_eol("Chironomus riparius"))
# same
as.eol(get_eol(c("Chironomus riparius","Pinus contorta")))
# numeric
as.eol(10247706)
# numeric vector, length > 1
as.eol(c(6985636,12188704,10247706))
# character
as.eol("6985636")
# character vector, length > 1
as.eol(c("6985636","12188704","10247706"))
# list, either numeric or character
as.eol(list("6985636","12188704","10247706"))
## dont check, much faster
as.eol("6985636", check=FALSE)
as.eol(6985636, check=FALSE)
as.eol(c("6985636","12188704","10247706"), check=FALSE)
as.eol(list("6985636","12188704","10247706"), check=FALSE)

(out <- as.eol(c(6985636,12188704,10247706)))
data.frame(out)
as.eol( data.frame(out) )

# Get all data back
get_eol_("Poa annua")
get_eol_("Poa annua", rows=2)
get_eol_("Poa annua", rows=1:2)
get_eol_(c("asdfadfasd", "Pinus contorta"))
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other taxonomic-ids: 
\code{\link{get_bold}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_itis}()},
\code{\link{get_iucn}()},
\code{\link{get_natserv}()},
\code{\link{get_nbn}()},
\code{\link{get_ncbi}()},
\code{\link{get_pow}()},
\code{\link{get_tol}()},
\code{\link{get_tps}()},
\code{\link{get_wiki}()},
\code{\link{get_worms}()},
\code{\link{id-accessors}}
}
\author{
Scott Chamberlain, \email{myrmecocystus@gmail.com}
}
\concept{taxonomic-ids}
