% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_genes}
\alias{get_genes}
\title{Retrieve gene sequences from NCBI by accession number.}
\usage{
get_genes(ids, format = "fasta", verbose = TRUE)
}
\arguments{
\item{ids}{GenBank ids to search for (character).}

\item{format}{Return type, e.g., "fasta"}

\item{verbose}{logical; If TRUE (default), informative messages printed.}
}
\description{
Function name changed to ncbi_getbyid.
}

