% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsn.R
\name{get_itis}
\alias{get_itis}
\alias{get_tsn}
\alias{as.itis}
\alias{as.itis.itis}
\alias{as.itis.character}
\alias{as.itis.list}
\alias{as.itis.numeric}
\alias{as.itis.data.frame}
\alias{get_itis_}
\alias{get_tsn_}
\title{Get the TSN code for a search term.}
\usage{
get_itis(
  sci_com,
  searchtype = "scientific",
  accepted = FALSE,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  searchterm = NULL,
  ...
)

get_tsn(
  sci_com,
  searchtype = "scientific",
  accepted = FALSE,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  searchterm = NULL,
  ...
)

as.itis(x, check = TRUE)

\method{as.itis}{itis}(x, check = TRUE)

\method{as.itis}{character}(x, check = TRUE)

\method{as.itis}{list}(x, check = TRUE)

\method{as.itis}{numeric}(x, check = TRUE)

\method{as.itis}{data.frame}(x, check = TRUE)

get_itis_(
  sci_com,
  messages = TRUE,
  searchtype = "scientific",
  accepted = TRUE,
  rows = NA,
  searchterm = NULL,
  ...
)

get_tsn_(
  sci_com,
  messages = TRUE,
  searchtype = "scientific",
  accepted = TRUE,
  rows = NA,
  searchterm = NULL,
  ...
)
}
\arguments{
\item{sci_com}{character; A vector of common or scientific names.
Or, a \code{taxon_state} object (see \link{taxon-state})}

\item{searchtype}{character; One of 'scientific' or 'common', or any
unique abbreviation}

\item{accepted}{logical; If TRUE, removes names that are not accepted valid
names by ITIS. Set to \code{FALSE} (default) to give back both accepted
and unaccepted names.}

\item{ask}{logical; should get_itis be run in interactive mode?
If \code{TRUE} and more than one TSN is found for the species, the user is
asked for input. If \code{FALSE} NA is returned for multiple matches.}

\item{messages}{logical; should progress be printed?}

\item{rows}{numeric; Any number from 1 to infinity. If the default NA, all
rows are considered. Note that this function still only gives back a tsn
class object with one to many identifiers. See
\code{\link[=get_itis_]{get_itis_()}} to get back all, or a subset, of the raw
data that you are presented during the ask process.}

\item{searchterm}{Deprecated, see \code{sci_com}}

\item{...}{Ignored}

\item{x}{Input to as.itis}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.itis]{as.itis()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Retrieve the taxonomic serial numbers (TSN) of a taxon from ITIS.
}
\examples{
\dontrun{
get_itis("Quercus douglasii")
get_itis("Chironomus riparius")
get_itis(c("Chironomus riparius","Quercus douglasii"))
splist <- c("annona cherimola", 'annona muricata', "quercus robur",
		"shorea robusta", "pandanus patina", "oryza sativa", "durio zibethinus")
get_itis(splist, messages=FALSE)

# specify rows to limit choices available
get_itis('Arni')
get_itis('Arni', rows=1)
get_itis('Arni', rows=1:2)

# When not found
get_itis("howdy")
get_itis(c("Chironomus riparius", "howdy"))

# Using common names
get_itis("black bear", searchtype="common")

# Convert a tsn without class information to a itis class
as.itis(get_itis("Quercus douglasii")) # already a itis, returns the same
as.itis(get_itis(c("Chironomus riparius","Pinus contorta"))) # same
as.itis(19322) # numeric
as.itis(c(19322,129313,506198)) # numeric vector, length > 1
as.itis("19322") # character
as.itis(c("19322","129313","506198")) # character vector, length > 1
as.itis(list("19322","129313","506198")) # list, either numeric or character
## dont check, much faster
as.itis("19322", check=FALSE)
as.itis(19322, check=FALSE)
as.itis(c("19322","129313","506198"), check=FALSE)
as.itis(list("19322","129313","506198"), check=FALSE)

(out <- as.itis(c(19322,129313,506198)))
data.frame(out)
as.itis( data.frame(out) )

# Get all data back
get_itis_("Arni")
get_itis_("Arni", rows=1)
get_itis_("Arni", rows=1:2)
get_itis_(c("asdfadfasd","Pinus contorta"), rows=1:5)
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other taxonomic-ids: 
\code{\link{get_bold}()},
\code{\link{get_eol}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_iucn}()},
\code{\link{get_natserv}()},
\code{\link{get_nbn}()},
\code{\link{get_ncbi}()},
\code{\link{get_pow}()},
\code{\link{get_tolid}()},
\code{\link{get_tps}()},
\code{\link{get_wiki}()},
\code{\link{get_worms}()}
}
\concept{taxonomic-ids}
