% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_natservid.R
\name{get_natserv}
\alias{get_natserv}
\alias{get_natservid}
\alias{as.natserv}
\alias{as.natserv.natserv}
\alias{as.natserv.character}
\alias{as.natserv.list}
\alias{as.natserv.numeric}
\alias{as.natserv.data.frame}
\alias{get_natserv_}
\alias{get_natservid_}
\title{Get NatureServe taxonomic ID for a taxon name}
\usage{
get_natserv(
  sci_com,
  searchtype = "scientific",
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  query = NULL,
  ...
)

get_natservid(
  sci_com,
  searchtype = "scientific",
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  query = NULL,
  ...
)

as.natserv(x, check = TRUE)

\method{as.natserv}{natserv}(x, check = TRUE)

\method{as.natserv}{character}(x, check = TRUE)

\method{as.natserv}{list}(x, check = TRUE)

\method{as.natserv}{numeric}(x, check = TRUE)

\method{as.natserv}{data.frame}(x, check = TRUE)

get_natserv_(
  sci_com,
  searchtype = "scientific",
  messages = TRUE,
  rows = NA,
  query = NULL,
  ...
)

get_natservid_(
  sci_com,
  searchtype = "scientific",
  messages = TRUE,
  rows = NA,
  query = NULL,
  ...
)
}
\arguments{
\item{sci_com}{character; A vector of common or scientific names. Or, a
\code{taxon_state} object (see \link{taxon-state})}

\item{searchtype}{character; One of 'scientific' (default) or 'common'.
This doesn't affect the query to NatureServe - but rather affects what
column of data is targeted in name filtering post data request.}

\item{ask}{logical; should get_natserv be run in interactive mode?
If \code{TRUE} and more than one wormsid is found for the species, the
user is asked for input. If \code{FALSE} NA is returned for
multiple matches. default: \code{TRUE}}

\item{messages}{logical; should progress be printed? default: \code{TRUE}}

\item{rows}{numeric; Any number from 1 to infinity. If the default NaN, all
rows are considered. Note that this function still only gives back a
natserv class object with one to many identifiers. See
\code{get_natserv_()} to get back all, or a subset, of the raw
data that you are presented during the ask process.}

\item{query}{Deprecated, see \code{sci_com}}

\item{...}{curl options passed on to \link[crul:verb-POST]{crul::verb-POST}}

\item{x}{Input to \code{as.natserv}}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.natserv]{as.natserv()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Get NatureServe taxonomic ID for a taxon name
}
\note{
Authentication no longer required
}
\examples{
\dontrun{
(x <- get_natserv("Helianthus annuus"))
library(taxa)
tax_name(x)
tax_id(x)

get_natserv('Gadus morhua')
get_natserv(c("Helianthus annuus", 'Gadus morhua'))

# specify rows to limit choices available
get_natserv('Ruby Quaker Moth', 'common')
get_natserv('Ruby*', 'common')
get_natserv('Ruby*', 'common', rows=1)
get_natserv('Ruby*', 'common', rows=1:2)

# When not found
get_natserv("howdy")
get_natserv(c('Gadus morhua', "howdy"))

# Convert a natserv without class information to a natserv class
# already a natserv, returns the same
as.natserv(get_natserv('Pomatomus saltatrix'))
# same
as.natserv(get_natserv(c('Gadus morhua', 'Pomatomus saltatrix')))
# character
as.natserv(101905)
# character vector, length > 1
as.natserv(c(101905, 101998))
# list, either numeric or character
as.natserv(list(101905, 101998))
## dont check, much faster
as.natserv(101905, check = FALSE)
as.natserv(c(101905, 101998), check = FALSE)
as.natserv(list(101905, 101998), check = FALSE)

(out <- as.natserv(c(101905, 101998), check = FALSE))
data.frame(out)
as.natserv( data.frame(out) )

# Get all data back
get_natserv_("Helianthus")
get_natserv_("Ruby*", searchtype = "common")
get_natserv_("Ruby*", searchtype = "common", rows=1:3)
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other taxonomic-ids: 
\code{\link{get_bold}()},
\code{\link{get_eol}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_iucn}()},
\code{\link{get_nbn}()},
\code{\link{get_pow}()},
\code{\link{get_tolid}()},
\code{\link{get_tps}()},
\code{\link{get_tsn}()},
\code{\link{get_uid}()},
\code{\link{get_wiki}()},
\code{\link{get_worms}()}
}
\concept{taxonomic-ids}
