% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_scamitid}
\alias{get_scamitid}
\title{Get the SCAMIT ID from taxonomic names.}
\usage{
get_scamitid(name, rank = "genus", ask = TRUE, verbose = TRUE)
}
\arguments{
\item{name}{The string to search for.}

\item{rank}{Taxonomic rank to search against. One of phylum, subphylum, class, subclass,
infraclass, superorder, order, suborder, infraorder, superfamily, family, subfamily,
tribe, genus (default), subgenus, or species.}

\item{ask}{logical; should get_scamitid be run in interactive mode?
If TRUE and more than one ID is found for the species, the user is asked for
input. If FALSE NA is returned for multiple matches.}

\item{verbose}{logical; If TRUE the actual taxon queried is printed on the console.}
}
\description{
Get the SCAMIT ID from taxonomic names.
}
\examples{
get_scamitid(name='Cliona', rank='genus')
}

