% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tolid.R
\name{get_tol}
\alias{get_tol}
\alias{get_tolid}
\alias{as.tol}
\alias{as.tol.tol}
\alias{as.tol.character}
\alias{as.tol.list}
\alias{as.tol.numeric}
\alias{as.tol.data.frame}
\alias{get_tol_}
\alias{get_tolid_}
\title{Get the OTT id for a search term}
\usage{
get_tol(sci, ask = TRUE, messages = TRUE, rows = NA, sciname = NULL, ...)

get_tolid(sci, ask = TRUE, messages = TRUE, rows = NA, sciname = NULL, ...)

as.tol(x, check = TRUE)

\method{as.tol}{tol}(x, check = TRUE)

\method{as.tol}{character}(x, check = TRUE)

\method{as.tol}{list}(x, check = TRUE)

\method{as.tol}{numeric}(x, check = TRUE)

\method{as.tol}{data.frame}(x, check = TRUE)

get_tol_(sci, messages = TRUE, rows = NA, sciname = NULL)

get_tolid_(sci, messages = TRUE, rows = NA, sciname = NULL)
}
\arguments{
\item{sci}{character; one or more scientific names. Or, a \code{taxon_state}
object (see \link{taxon-state})}

\item{ask}{logical; should \code{get_tol} be run in interactive mode?
If \code{TRUE} and more than one TOL is found for the species, the user is
asked for input. If \code{FALSE} NA is returned for multiple matches.}

\item{messages}{logical; should progress be printed?}

\item{rows}{numeric; Any number from 1 to infinity. If the default NA, all
rows are considered. Note that this function still only gives back a tol
class object with one to many identifiers. See \code{\link[=get_tol_]{get_tol_()}}
to get back all, or a subset, of the raw data that you are presented during
the ask process.}

\item{sciname}{Deprecated, see \code{sci}}

\item{...}{Ignored}

\item{x}{Input to \code{as.tol}}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.tol]{as.tol()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Retrieve the Open Tree of Life Taxonomy (OTT) id of a taxon from
OpenTreeOfLife
}
\examples{
\dontrun{
get_tol(sci = "Quercus douglasii")
get_tol(sci = "Chironomus riparius")
get_tol(c("Chironomus riparius","Quercus douglasii"))
splist <- c("annona cherimola", 'annona muricata', "quercus robur",
		"shorea robusta", "pandanus patina", "oryza sativa", "durio zibethinus")
get_tol(splist, messages=FALSE)

# specify rows to limit choices available
get_tol('Arni')
get_tol('Arni', rows=1)
get_tol('Arni', rows=1:2)

# When not found
get_tol("howdy")
get_tol(c("Chironomus riparius", "howdy"))

# Convert a tol without class information to a tol class
as.tol(get_tol("Quercus douglasii")) # already a tol, returns the same
as.tol(get_tol(c("Chironomus riparius","Pinus contorta"))) # same
as.tol(5907893) # numeric
as.tol(c(3930798,515712,872577)) # numeric vector, length > 1
as.tol("3930798") # character
as.tol(c("3930798","515712","872577")) # character vector, length > 1
as.tol(list("3930798","515712","872577")) # list, either numeric or character
## dont check, much faster
as.tol("3930798", check=FALSE)
as.tol(3930798, check=FALSE)
as.tol(c("3930798","515712","872577"), check=FALSE)
as.tol(list("3930798","515712","872577"), check=FALSE)

(out <- as.tol(c(3930798,515712,872577)))
data.frame(out)
as.tol( data.frame(out) )

# Get all data back
get_tol_("Arni")
get_tol_("Arni", rows=1)
get_tol_("Arni", rows=1:2)
get_tol_(c("asdfadfasd","Pinus contorta"))
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other taxonomic-ids: 
\code{\link{get_bold}()},
\code{\link{get_eol}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_itis}()},
\code{\link{get_iucn}()},
\code{\link{get_natserv}()},
\code{\link{get_nbn}()},
\code{\link{get_ncbi}()},
\code{\link{get_pow}()},
\code{\link{get_tps}()},
\code{\link{get_wiki}()},
\code{\link{get_worms}()}
}
\concept{taxonomic-ids}
