% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ubioid.R
\name{get_ubioid}
\alias{get_ubioid}
\alias{as.ubioid}
\alias{get_ubioid_}
\title{Get the uBio id for a search term}
\usage{
get_ubioid(...)

as.ubioid(...)

get_ubioid_(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\keyword{internal}
