% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_ubioid}
\alias{as.ubioid}
\alias{as.ubioid.character}
\alias{as.ubioid.list}
\alias{as.ubioid.numeric}
\alias{as.ubioid.ubioid}
\alias{get_ubioid}
\title{Get the uBio id for a search term.}
\usage{
get_ubioid(searchterm, searchtype = "scientific", ask = TRUE,
  verbose = TRUE)

as.ubioid(x)

\method{as.ubioid}{ubioid}(x)

\method{as.ubioid}{character}(x)

\method{as.ubioid}{list}(x)

\method{as.ubioid}{numeric}(x)
}
\arguments{
\item{searchterm}{character; A vector of common or scientific names.}

\item{searchtype}{character; One of 'scientific' or 'common', or any unique abbreviation}

\item{ask}{logical; should get_tsn be run in interactive mode?
If TRUE and more than one TSN is found for teh species, the user is asked for
input. If FALSE NA is returned for multiple matches.}

\item{verbose}{logical; should progress be printed?}

\item{x}{Input to \code{\link{as.ubioid}}}
}
\value{
A vector of uBio ids. If a taxon is not found NA is given. If more than one uBio
   id is found the function asks for user input (if ask = TRUE), otherwise returns NA.
   Comes with an attribute \emph{match} to investigate the reason for NA (either 'not found',
   'found' or if ask = FALSE 'multi match')
}
\description{
Retrieve the uBio id of a taxon. This function uses \code{\link[taxize]{ubio_search}} internally
to search for names.
}
\examples{
\donttest{
get_ubioid(searchterm = "Astragalus aduncus")
get_ubioid(c("Salvelinus fontinalis","Pomacentrus brachialis"))
splist <- c("Salvelinus fontinalis", 'Pomacentrus brachialis', "Leptocottus armatus",
		"Clinocottus recalvus", "Trachurus trachurus", "Harengula clupeola")
get_ubioid(splist, verbose=FALSE)

# When not found
get_ubioid(searchterm="howdy")
get_ubioid(c("Salvelinus fontinalis", "howdy"))

# Using common names
get_ubioid(searchterm="great white shark", searchtype="common")
get_ubioid(searchterm=c("bull shark", "whale shark"), searchtype="common")

# Convert a ubioid without class information to a ubioid class
as.ubioid(get_ubioid("Astragalus aduncus")) # already a ubioid, returns the same
as.ubioid(get_ubioid(c("Chironomus riparius","Pinus contorta"))) # same
as.ubioid(2843601) # numeric
as.ubioid(c(2843601,3339,9696)) # numeric vector, length > 1
as.ubioid("2843601") # character
as.ubioid(c("2843601","3339","9696")) # character vector, length > 1
as.ubioid(list("2843601","3339","9696")) # list, either numeric or character
}
}
\seealso{
\code{\link[taxize]{get_uid}}, \code{\link[taxize]{ubio_search}}
}

