% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wormsid.R
\name{get_worms}
\alias{get_worms}
\alias{get_wormsid}
\alias{as.worms}
\alias{as.worms.worms}
\alias{as.worms.character}
\alias{as.worms.list}
\alias{as.worms.numeric}
\alias{as.worms.data.frame}
\alias{get_worms_}
\alias{get_wormsid_}
\title{Get Worms ID for a taxon name}
\usage{
get_worms(
  sci_com,
  searchtype = "scientific",
  marine_only = TRUE,
  fuzzy = NULL,
  accepted = FALSE,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  query = NULL,
  ...
)

get_wormsid(
  sci_com,
  searchtype = "scientific",
  marine_only = TRUE,
  fuzzy = NULL,
  accepted = FALSE,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  query = NULL,
  ...
)

as.worms(x, check = TRUE)

\method{as.worms}{worms}(x, check = TRUE)

\method{as.worms}{character}(x, check = TRUE)

\method{as.worms}{list}(x, check = TRUE)

\method{as.worms}{numeric}(x, check = TRUE)

\method{as.worms}{data.frame}(x, check = TRUE)

get_worms_(
  sci_com,
  messages = TRUE,
  searchtype = "scientific",
  marine_only = TRUE,
  fuzzy = NULL,
  accepted = TRUE,
  rows = NA,
  query = NULL,
  ...
)

get_wormsid_(
  sci_com,
  messages = TRUE,
  searchtype = "scientific",
  marine_only = TRUE,
  fuzzy = NULL,
  accepted = TRUE,
  rows = NA,
  query = NULL,
  ...
)
}
\arguments{
\item{sci_com}{character; A vector of common or scientific names. Or, a
\code{taxon_state} object (see \link{taxon-state})}

\item{searchtype}{character; One of 'scientific' or 'common', or any unique
abbreviation}

\item{marine_only}{logical; marine only? default: \code{TRUE} (only used
when \code{searchtype="scientific"}); passed on to \code{\link[worrms:wm_records_name]{worrms::wm_records_name()}}}

\item{fuzzy}{logical; fuzzy search. default: \code{NULL} (\code{TRUE} for
\code{searchtype="scientific"} and \code{FALSE} for \code{searchtype="common"} to match
the default values for those parameters in \pkg{worrms} package); passed on
to \code{\link[worrms:wm_records_name]{worrms::wm_records_name()}} or \code{\link[worrms:wm_records_common]{worrms::wm_records_common()}}}

\item{accepted}{logical; If TRUE, removes names that are not accepted valid
names by WORMS. Set to \code{FALSE} (default) to give back both accepted
and unaccepted names.}

\item{ask}{logical; should get_worms be run in interactive mode?
If \code{TRUE} and more than one worms is found for the species, the
user is asked for input. If \code{FALSE} NA is returned for
multiple matches.}

\item{messages}{logical; should progress be printed?}

\item{rows}{numeric; Any number from 1 to infinity. If the default NaN, all
rows are considered. Note that this function still only gives back a worms
class object with one to many identifiers. See \code{\link[=get_worms_]{get_worms_()}} to get back
all, or a subset, of the raw data that you are presented during the ask
process.}

\item{query}{Deprecated, see \code{sci_com}}

\item{...}{Ignored}

\item{x}{Input to \code{as.worms}}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.worms]{as.worms()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Retrieve Worms ID of a taxon from World Register of Marine
Species (WORMS).
}
\examples{
\dontrun{
(x <- get_worms('Gadus morhua'))
data.frame(x)

get_worms('Pomatomus saltatrix')
get_worms(c("Gadus morhua", "Lichenopora neapolitana"))

# marine_only
get_worms("Apedinella", marine_only=TRUE)
get_worms("Apedinella", marine_only=FALSE)

# fuzzy
## searchtype="scientific": fuzzy is TRUE by default
get_worms("Platypro", searchtype="scientific", fuzzy=TRUE)
get_worms("Platypro", searchtype="scientific", fuzzy=FALSE)
## searchtype="common": fuzzy is FALSE by default
get_worms("clam", searchtype="common", fuzzy=FALSE)
get_worms("clam", searchtype="common", fuzzy=TRUE)

# by common name
get_worms("dolphin", 'common')
get_worms("clam", 'common')

# specify rows to limit choices available
get_worms('Plat')
get_worms('Plat', rows=1)
get_worms('Plat', rows=1:2)

# When not found
get_worms("howdy")
get_worms(c('Gadus morhua', "howdy"))

# Convert a worms without class information to a worms class
# already a worms, returns the same
as.worms(get_worms('Gadus morhua'))
# same
as.worms(get_worms(c('Gadus morhua', 'Pomatomus saltatrix')))
# numeric
as.worms(126436)
# numeric vector, length > 1
as.worms(c(126436,151482))
# character
as.worms("126436")
# character vector, length > 1
as.worms(c("126436","151482"))
# list, either numeric or character
as.worms(list("126436","151482"))
## dont check, much faster
as.worms("126436", check=FALSE)
as.worms(126436, check=FALSE)
as.worms(c("126436","151482"), check=FALSE)
as.worms(list("126436","151482"), check=FALSE)

(out <- as.worms(c(126436,151482)))
data.frame(out)
as.worms( data.frame(out) )

# Get all data back
get_worms_("Plat")
get_worms_("Plat", rows=1)
get_worms_("Plat", rows=1:2)
get_worms_("Plat", rows=1:75)
# get_worms_(c("asdfadfasd","Plat"), rows=1:5)
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other taxonomic-ids: 
\code{\link{get_bold}()},
\code{\link{get_eol}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_itis}()},
\code{\link{get_iucn}()},
\code{\link{get_natserv}()},
\code{\link{get_nbn}()},
\code{\link{get_ncbi}()},
\code{\link{get_pow}()},
\code{\link{get_tol}()},
\code{\link{get_tps}()},
\code{\link{get_wiki}()},
\code{\link{id-accessors}}
}
\concept{taxonomic-ids}
