% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcommentdetailfromtsn}
\alias{getcommentdetailfromtsn}
\title{Get comment detail from TSN}
\usage{
getcommentdetailfromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\value{
A data.frame with results.
}
\description{
Get comment detail from TSN
}
\examples{
\dontrun{
library("httr")
getcommentdetailfromtsn(tsn=180543, config=timeout(4))

# using sql locally
backend_get()
backend_set("local")
getcommentdetailfromtsn(tsn=180543)
}
}
\keyword{internal}

