% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcommonnamesfromtsn}
\alias{getcommonnamesfromtsn}
\title{Get common names from tsn}
\usage{
getcommonnamesfromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get common names from tsn
}
\examples{
\dontrun{
getcommonnamesfromtsn(183833, config=timeout(1))

# local
getcommonnamesfromtsn(183833, backend="local")
}
}
\keyword{internal}

