\name{getcoremetadatafromtsn}
\alias{getcoremetadatafromtsn}
\title{match count}
\usage{
  getcoremetadatafromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getCoreMetadataFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

