% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcoveragefromtsn}
\alias{getcoveragefromtsn}
\title{Get coverge from tsn}
\usage{
getcoveragefromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get coverge from tsn
}
\examples{
\dontrun{
getcoveragefromtsn(tsn=28727, config=timeout(4))  # coverage data
getcoveragefromtsn(526852, config=timeout(4))  # no coverage data

backend_set("local")
getcoveragefromtsn(tsn=28727)
}
}
\keyword{internal}

