% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcredibilityratingfromtsn}
\alias{getcredibilityratingfromtsn}
\title{Get credibility rating from tsn}
\usage{
getcredibilityratingfromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get credibility rating from tsn
}
\examples{
\dontrun{
getcredibilityratingfromtsn(526852, config=timeout(4))

backend_set("local")
getcredibilityratingfromtsn(526852)
}
}
\keyword{internal}

