% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcredibilityratings}
\alias{getcredibilityratings}
\title{Get possible credibility ratings}
\usage{
getcredibilityratings(backend = NULL, ...)
}
\arguments{
\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible credibility ratings
}
\examples{
\dontrun{
getcredibilityratings(config=timeout(3))

backend_set("local")
getcredibilityratings()
}
}
\keyword{internal}

