% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getdatedatafromtsn}
\alias{getdatedatafromtsn}
\title{Get date data from tsn}
\usage{
getdatedatafromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get date data from tsn
}
\examples{
\dontrun{
getdatedatafromtsn(180543, config=timeout(3))

backend_set("localsql")
getdatedatafromtsn(180543)
}
}
\keyword{internal}

