% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getdescription}
\alias{getdescription}
\title{Get description of the ITIS service}
\usage{
getdescription(...)
}
\description{
Get description of the ITIS service
}
\examples{
\donttest{
getdescription(config=timeout(1))
}
}
\keyword{internal}

