\name{getgeographicdivisionsfromtsn}
\alias{getgeographicdivisionsfromtsn}
\title{match count}
\usage{
  getgeographicdivisionsfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getGeographicDivisionsFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

