% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getgeographicvalues}
\alias{getgeographicvalues}
\title{Get all possible geographic values}
\usage{
getgeographicvalues(backend = NULL, ...)
}
\arguments{
\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get all possible geographic values
}
\examples{
\dontrun{
getgeographicvalues("api")

backend_set("local")
getgeographicvalues()
}
}
\keyword{internal}

