% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getglobalspeciescompletenessfromtsn}
\alias{getglobalspeciescompletenessfromtsn}
\title{Get global species completeness from tsn}
\usage{
getglobalspeciescompletenessfromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get global species completeness from tsn
}
\examples{
\dontrun{
getglobalspeciescompletenessfromtsn(180541, "api")

backend_set("local")
getglobalspeciescompletenessfromtsn(180541)
}
}
\keyword{internal}

