% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{gethierarchydownfromtsn}
\alias{gethierarchydownfromtsn}
\title{Get hierarchy down from tsn}
\usage{
gethierarchydownfromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get hierarchy down from tsn
}
\examples{
\dontrun{
gethierarchydownfromtsn(161030, config=timeout(3))

backend_set("local")
gethierarchydownfromtsn(161030)
}
}
\keyword{internal}

