% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getitistermsfromcommonname}
\alias{getitistermsfromcommonname}
\title{Get itis terms from common names}
\usage{
getitistermsfromcommonname(x, backend = NULL, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get itis terms from common names
}
\examples{
\dontrun{
getitistermsfromcommonname("buya")

library('httr')
getitistermsfromcommonname("buya", config=timeout(1))

backend_set("local")
getitistermsfromcommonname(x="buya")
}
}
\keyword{internal}

