% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getitistermsfromscientificname}
\alias{getitistermsfromscientificname}
\title{Get itis terms from scientific names}
\usage{
getitistermsfromscientificname(x, backend = NULL, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get itis terms from scientific names
}
\examples{
\dontrun{
getitistermsfromscientificname("ursidae")
getitistermsfromscientificname("Ursus")

backend_set("local")
getitistermsfromscientificname("ursidae")
getitistermsfromscientificname("Ursus")
getitistermsfromscientificname("Poa")
}
}
\keyword{internal}

