% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getjurisdictionoriginvalues}
\alias{getjurisdictionoriginvalues}
\title{Get jurisdiction origin values}
\usage{
getjurisdictionoriginvalues(backend = NULL, ...)
}
\arguments{
\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get jurisdiction origin values
}
\examples{
\dontrun{
getjurisdictionoriginvalues(config=timeout(3))

backend_set("localsql")
getjurisdictionoriginvalues()
}
}
\keyword{internal}

