% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getjurisdictionvalues}
\alias{getjurisdictionvalues}
\title{Get possible jurisdiction values}
\usage{
getjurisdictionvalues(backend = NULL, ...)
}
\arguments{
\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible jurisdiction values
}
\examples{
\dontrun{
getjurisdictionvalues(config=timeout(3))

backend_set("localsql")
getjurisdictionvalues()
}
}
\keyword{internal}

