% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getkingdomnamefromtsn}
\alias{getkingdomnamefromtsn}
\title{Get kingdom names from tsn}
\usage{
getkingdomnamefromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get kingdom names from tsn
}
\examples{
\dontrun{
getkingdomnamefromtsn(202385, config=timeout(3))

backend_set("localsql")
getkingdomnamefromtsn(202385)
}
}
\keyword{internal}

