% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getkingdomnames}
\alias{getkingdomnames}
\title{Get all possible kingdom names}
\usage{
getkingdomnames(backend = NULL, ...)
}
\arguments{
\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get all possible kingdom names
}
\examples{
\dontrun{
getkingdomnames(config=timeout(3))

backend_set("local")
getkingdomnames()
}
}
\keyword{internal}

