% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getlastchangedate}
\alias{getlastchangedate}
\title{Provides the date the ITIS database was last updated.}
\usage{
getlastchangedate(backend = NULL, ...)
}
\arguments{
\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Provides the date the ITIS database was last updated.
}
\examples{
\dontrun{
getlastchangedate("api")
getlastchangedate("local")
}
}
\keyword{internal}

