% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getlsidfromtsn}
\alias{getlsidfromtsn}
\title{Gets the unique LSID for the TSN, or an empty result if there is no match.}
\usage{
getlsidfromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle())
}
\description{
Gets the unique LSID for the TSN, or an empty result if there is no match.
}
\examples{
\donttest{
getlsidfromtsn(tsn = 155166)
}
}
\keyword{internal}

