% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getpublicationsfromtsn}
\alias{getpublicationsfromtsn}
\title{Returns a list of the pulications used for the TSN.}
\usage{
getpublicationsfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Returns a list of the pulications used for the TSN.
}
\examples{
\donttest{
getpublicationsfromtsn(tsn = 70340)
}
}
\keyword{internal}

