% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getrecordfromlsid}
\alias{getrecordfromlsid}
\title{Gets the partial ITIS record for the TSN in the LSID, found by comparing the
 TSN in the search key to the TSN field. Returns an empty result set if
 there is no match or the TSN is invalid.}
\usage{
getrecordfromlsid(lsid = NA, curlopts = list(), verbose = TRUE)
}
\arguments{
\item{lsid}{lsid for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{verbose}{(logical) Print messages or not. Default to TRUE.}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Gets the partial ITIS record for the TSN in the LSID, found by comparing the
 TSN in the search key to the TSN field. Returns an empty result set if
 there is no match or the TSN is invalid.
}
\examples{
\donttest{
getrecordfromlsid(lsid = "urn:lsid:itis.gov:itis_tsn:180543")
}
}
\keyword{internal}

