% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getscamithierarchy}
\alias{getscamithierarchy}
\title{Get scamit hierarchy}
\usage{
getscamithierarchy(id)
}
\arguments{
\item{id}{Scamit identifier}
}
\description{
Get scamit hierarchy
}
\examples{
getscamithierarchy(2522)
}
\keyword{internal}

