% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getscientificnamefromtsn}
\alias{getscientificnamefromtsn}
\title{Returns the scientific name for the TSN. Also returns the component parts
   (names and indicators) of the scientific name.}
\usage{
getscientificnamefromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the scientific name for the TSN. Also returns the component parts
   (names and indicators) of the scientific name.
}
\examples{
\dontrun{
getscientificnamefromtsn(531894, "api")
getscientificnamefromtsn(531894, "local")
}
}
\keyword{internal}

