% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getsynonymnamesfromtsn}
\alias{getsynonymnamesfromtsn}
\title{Returns a list of the synonyms (if any) for the TSN.}
\usage{
getsynonymnamesfromtsn(tsn, backend = NULL, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns a list of the synonyms (if any) for the TSN.
}
\examples{
\dontrun{
getsynonymnamesfromtsn(183671, "api") # tsn not accepted
getsynonymnamesfromtsn(526852, "api") # tsn accepted

getsynonymnamesfromtsn(183671, "local") # tsn not accepted
getsynonymnamesfromtsn(526852, "local") # tsn accepted
}
}
\keyword{internal}

