% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettaxonomicusagefromtsn}
\alias{gettaxonomicusagefromtsn}
\title{Returns the usage information for the TSN.}
\usage{
gettaxonomicusagefromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the usage information for the TSN.
}
\examples{
\dontrun{
gettaxonomicusagefromtsn(526852, config=timeout(3))
}
}
\keyword{internal}

