% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{gettsnbyvernacularlanguage}
\alias{gettsnbyvernacularlanguage}
\title{Get tsn by vernacular language}
\usage{
gettsnbyvernacularlanguage(language, backend = NULL, ...)
}
\arguments{
\item{language}{A string containing the language. This is a language string,
not the international language code (character)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get tsn by vernacular language
}
\examples{
\dontrun{
gettsnbyvernacularlanguage("french", "api")
gettsnbyvernacularlanguage("french", "local")
}
}
\keyword{internal}

