% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettsnfromlsid}
\alias{gettsnfromlsid}
\title{Gets the TSN corresponding to the LSID, or an empty result if there is no match.}
\usage{
gettsnfromlsid(lsid, ...)
}
\arguments{
\item{lsid}{lsid for a taxonomic group (character)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Gets the TSN corresponding to the LSID, or an empty result if there is no match.
}
\examples{
\dontrun{
gettsnfromlsid(lsid="urn:lsid:itis.gov:itis_tsn:28726", config=timeout(3))
gettsnfromlsid("urn:lsid:itis.gov:itis_tsn:0", config=timeout(3))
}
}
\keyword{internal}

