% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getunacceptabilityreasonfromtsn}
\alias{getunacceptabilityreasonfromtsn}
\title{Returns the unacceptability reason, if any, for the TSN.}
\usage{
getunacceptabilityreasonfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle())
}
\description{
Returns the unacceptability reason, if any, for the TSN.
}
\examples{
\donttest{
getunacceptabilityreasonfromtsn(tsn = 183671)
}
}
\keyword{internal}

