\name{getxmldata}
\alias{getxmldata}
\title{Function to parse xml data and decode strings}
\usage{
  getxmldata(obj, node, todecode)

  getxmldata(obj, node, todecode)
}
\description{
  Function to parse xml data and decode strings

  Function to parse xml data and decode strings
}
\keyword{internal}

