% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisd_isinvasive.R
\name{gisd_isinvasive}
\alias{gisd_isinvasive}
\title{Check invasive species status for a set of species from GISD database}
\usage{
gisd_isinvasive(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\keyword{internal}
