% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gna_data_sources.R
\name{gna_data_sources}
\alias{gna_data_sources}
\title{Get metadata about GNA data sources}
\usage{
gna_data_sources(output_type = "table", ...)
}
\arguments{
\item{output_type}{What format of output to return. Either \code{'json'},
\code{'list'}, or \code{'table'}.}

\item{...}{Passed to \link[crul:HttpClient]{crul::HttpClient}.}
}
\description{
Downloads metadata about Global Names Architecture (GNA) data sources
available to be used in other GNA functions.
}
\examples{
\dontrun{

gna_data_sources()
}

}
\author{
Zachary S.L. Foster
}
