% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_taxon.R
\name{id-accessors}
\alias{id-accessors}
\alias{txz_uri}
\alias{txz_match}
\alias{txz_mm}
\alias{txz_pm}
\alias{txz_misc}
\title{Access various metadata from taxonomic id objects}
\usage{
txz_uri(x)

txz_match(x)

txz_mm(x)

txz_pm(x)

txz_misc(x)
}
\arguments{
\item{x}{result of a call to a \verb{get_*} function, an object
of class \code{txid}}
}
\description{
Access various metadata from taxonomic id objects
}
\seealso{
Other taxonomic-ids: 
\code{\link{get_bold}()},
\code{\link{get_eol}()},
\code{\link{get_gbif}()},
\code{\link{get_ids}()},
\code{\link{get_itis}()},
\code{\link{get_iucn}()},
\code{\link{get_natserv}()},
\code{\link{get_nbn}()},
\code{\link{get_ncbi}()},
\code{\link{get_pow}()},
\code{\link{get_tol}()},
\code{\link{get_tps}()},
\code{\link{get_wiki}()},
\code{\link{get_worms}()}
}
\concept{taxonomic-ids}
