% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2name.R
\name{id2name}
\alias{id2name}
\alias{id2name.default}
\alias{id2name.tolid}
\alias{id2name.tsn}
\title{Taxonomic IDs to taxonomic names}
\usage{
id2name(x, db = NULL, ...)

\method{id2name}{default}(x, db = NULL, ...)

\method{id2name}{tolid}(x, ...)

\method{id2name}{tsn}(x, ...)
}
\arguments{
\item{x}{vector of taxonomic IDs (character or numeric)}

\item{db}{character; database to query. One or more of \code{tol},
or \code{itis}. Note that each taxonomic data source has their own 
identifiers, so that if you provide the wrong
\code{db} value for the identifier you could get a result, but it will
likely be wrong (not what you were expecting).}

\item{...}{Further args passed on to \code{tol_id2name} or
\code{\link{itis_getrecord}}
See those functions for what parameters can be passed on.}
}
\value{
A named list of data.frames
}
\description{
Taxonomic IDs to taxonomic names
}
\examples{
\dontrun{
id2name(19322, db = "itis")

# with TOL, get NCBI ID and pass to classification()
x <- id2name(515698, db = "tol")
classification(as.uid(x[[1]]$tax_sources_ncbi))
}
}
