\name{ifelsedata}
\alias{ifelsedata}
\title{Function to parse xml data, decode strings, and return NULL if logical is zero.}
\usage{
  ifelsedata(x, y, z)

  ifelsedata(x, y, z)
}
\description{
  Function to parse xml data, decode strings, and return
  NULL if logical is zero.

  Function to parse xml data, decode strings, and return
  NULL if logical is zero.
}
\keyword{internal}

