% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion.R
\name{ion}
\alias{ion}
\title{ION - Index to Organism Names}
\usage{
ion(x, ...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{lsid}{An LSID number. Required.}
}
\value{
A data.frame
}
\description{
ION - Index to Organism Names
}
\examples{
\dontrun{
ion(155166)
ion(298678)
ion(4796748) # ursus americanus
ion(1280626) # puma concolor
}
}
\references{
\url{http://www.organismnames.com/}
}

