% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_getrecord.R
\name{itis_getrecord}
\alias{itis_getrecord}
\title{Get full ITIS record for one or more ITIS TSN's or lsid's.}
\usage{
itis_getrecord(values, by = "tsn", ...)
}
\arguments{
\item{values}{(character) One or more TSN's (taxonomic serial number) or lsid's for a
taxonomic group}

\item{by}{(character) By "tsn" (default) or "lsid"}

\item{...}{Further arguments passed on to \code{\link[ritis:full_record]{ritis::full_record}}}
}
\description{
Get full ITIS record for one or more ITIS TSN's or lsid's.
}
\details{
You can only enter values in tsn parameter or lsid, not both.
}
\examples{
\dontrun{
# by TSN
itis_getrecord(202385)
itis_getrecord(c(202385,70340))

# by lsid
itis_getrecord("urn:lsid:itis.gov:itis_tsn:202385", "lsid")
}
}
