% Generated by roxygen2 (4.0.1): do not edit by hand
\name{itis_ping}
\alias{itis_ping}
\title{Ping the ITIS API to see if it's working.}
\usage{
itis_ping()
}
\description{
And provides number of scientific and common names in a string.
}
\examples{
\donttest{
itis_ping()
}
}
\seealso{
\code{\link{eol_ping}}
}

