\name{iucn_status}
\alias{iucn_status}
\alias{iucn_status.iucn}
\title{Extractor Functions for iucn-class}
\usage{
  iucn_status(x, ...)

  \method{iucn_status}{iucn} (x, ...)
}
\arguments{
  \item{x}{iucn-object as returned by \code{iucn_summary}}

  \item{...}{Currently not used}
}
\value{
  A character vector with the status
}
\description{
  Extractor Functions for iucn-class
}

