% Generated by roxygen2 (4.0.1): do not edit by hand
\name{myco_search}
\alias{myco_search}
\title{Search Mycobank}
\usage{
myco_search(layout = "mbwservice", filter = "", limit = 100, ...)
}
\arguments{
\item{layout}{(character) One of 'mbwservice' or 'mycobank'. There is no explanation of what
these different layouts are. Help anyone?}

\item{filter}{(character) A variety of inputs. See Details.}

\item{limit}{(integer) Number to return. Default: 100}

\item{...}{Futher args passed onto httr::GET}
}
\value{
A list of length two, with slots for data and links, in which NROW(data) should
equal length(links).
}
\description{
Search Mycobank
}
\details{
For the filter p?arameter, use the following syntax:
\code{<FieldCode1><Operator>"<Value1>" AND|OR|NOT <FieldCode2><Operator>"<Value2>"}. Use the
special value: $NULL$ to filter on Null values and $EMPTY$ to filter on empty string values.
These two special values are not case sensitive.

For mbwservice, the options are

\itemize{
 \item Authors_ text, e.g., \code{Authors_ CONTAINS "Value"}
 \item Epithet_ text, e.g., \code{Epithet_ CONTAINS "Value"}
 \item MycoBankNr_ numeric, e.g., code{MycoBankNr_="1"}
 \item Name text, e.g., \code{Name CONTAINS "Value"}
 \item Year of publication DateTime (YYYYMMDDHHmmss), e.g., \code{NameYear_<"20140826022700"}
}

For mycobank, the options are

\itemize{
 \item Authors_ text, e.g., \code{Authors_ CONTAINS "Value"}
 \item Classification_ text, e.g., \code{Classification_="Value"}
 \item Epithet_ text, e.g., \code{Epithet_ CONTAINS "Value"}
 \item MycoBankNr_ numeric, e.g., code{MycoBankNr_="1"}
 \item Name text, e.g., \code{NAME CONTAINS "Value"}
 \item NameStatus_ text, e.g., \code{NameStatus_ = "Value"}
 \item Summary text, e.g., \code{E3787 CONTAINS "Value"}
 \item Synonymy text, e.g., \code{E4060 CONTAINS "Value"}
 \item Year of publication DateTime (YYYYMMDDHHmmss), e.g., \code{NameYear_<"20140826022700"}
}

Operator options

\itemize{
 \item =	Equality
 \item <	Lower than
 \item >	Greater than
 \item <> Diffirent
 \item <= Lower than or equals
 \item >= Greater than or equals
 \item CONTAINS	Contains the specified string
 \item STARTSWITH	Starts with the specified string
 \item ENDSWITH	Ends with the specified string
}
}
\examples{
\dontrun{
myco_search(filter='MycoBankNr_="344025"')
head(myco_search(filter='MycoBankNr_>"344025"'))
myco_search(filter='Name CONTAINS "Candida boidinii"')
myco_search(filter='Name CONTAINS "Candida"')
myco_search(filter='Name STARTSWITH "Candida"')
myco_search(filter='Name ENDSWITH "Candida"')

# using layout = mycobank
myco_search(layout='mycobank', filter='MycoBankNr_="344025"')
head(myco_search(layout='mycobank', filter='MycoBankNr_>"344025"'))

# Curl debugging
library('httr')
myco_search(filter='MycoBankNr_="344025"', config=verbose())
}
}
\references{
\url{http://www.mycobank.org/Services/Generic/Help.aspx?s=searchservice}
}

