% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_getbyid.R
\name{ncbi_getbyid}
\alias{ncbi_getbyid}
\title{Retrieve gene sequences from NCBI by accession number.}
\usage{
ncbi_getbyid(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{internal}
