% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_getbyname.R
\name{ncbi_getbyname}
\alias{ncbi_getbyname}
\title{Retrieve gene sequences from NCBI by taxon name and gene names.}
\usage{
ncbi_getbyname(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{internal}
