% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ncbi_search}
\alias{ncbi_search}
\title{Search for gene sequences available for a species from NCBI.}
\usage{
ncbi_search(taxa = NULL, id = NULL, seqrange = "1:3000",
  getrelated = FALSE, limit = 500, verbose = TRUE)
}
\arguments{
\item{taxa}{(character) Scientific name to search for.}

\item{id}{(character) Taxonomic id to search for. Not compatible with argument \code{taxa}.}

\item{seqrange}{(character) Sequence range, as e.g., \code{"1:1000"}. This is the range of
sequence lengths to search for. So \code{"1:1000"} means search for sequences from 1 to 1000
characters in length.}

\item{getrelated}{(logical) If \code{TRUE}, gets the longest sequences of a species
in the same genus as the one searched for. If \code{FALSE}, returns nothing if no match
found.}

\item{limit}{Number of sequences to search for and return. Max of 10,000.
If you search for 6000 records, and only 5000 are found, you will of course
only get 5000 back.}

\item{verbose}{(logical) If \code{TRUE} (default), informative messages printed.}
}
\value{
Data.frame of results.
}
\description{
Search for gene sequences available for a species from NCBI.
}
\details{
Removes predicted sequences so you don't have to remove them.
    Predicted sequences are those with accession numbers that have "XM_" or
		"XR_" prefixes.
}
\examples{
\donttest{
# A single species
out <- ncbi_search(taxa="Umbra limi", seqrange = "1:2000")
# Get the same species information using a taxonomy id
out <- ncbi_search(id = "75935", seqrange = "1:2000")
# If the taxon name is unique, using the taxon name and id are equivalent
all(ncbi_search(id = "75935") ==  ncbi_search(taxa="Umbra limi"))
# If the taxon name is not unique, use taxon id
#  "266948" is the uid for the butterfly genus, but there is also a genus of orchids with the
#  same name
nrow(ncbi_search(id = "266948")) ==  nrow(ncbi_search(taxa="Satyrium"))
# get list of genes available, removing non-unique
unique(out$gene_desc)
# does the string 'RAG1' exist in any of the gene names
out[grep("RAG1", out$gene_desc, ignore.case=TRUE),]

# A single species without records in NCBI
out <- ncbi_search(taxa="Sequoia wellingtonia", seqrange="1:2000", getrelated=TRUE)

# Many species, can run in parallel or not using plyr
species <- c("Salvelinus alpinus","Ictalurus nebulosus","Carassius auratus")
out2 <- ncbi_search(taxa=species, seqrange = "1:2000")
lapply(out2, head) # see heads of all
library("plyr")
out2df <- ldply(out2) # make data.frame of all
unique(out2df$gene_desc) # get list of genes available, removing non-unique
out2df[grep("60S ribosomal protein", out2df$gene_desc, ignore.case=TRUE),] # search across all
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}, Zachary Foster
}
\seealso{
\code{\link[taxize]{ncbi_getbyid}}, \code{\link[taxize]{ncbi_getbyname}}
}

