\name{\%notin\%}
\alias{\%notin\%}
\title{opposite of \%in\% : 'notin'}
\usage{
  x \%notin\% table
}
\arguments{
  \item{x}{the values to be matched.}

  \item{table}{the values to be matched against.}
}
\value{
  A logical vector of the same length as x.
}
\description{
  opposite of \%in\% : 'notin'
}
\keyword{internal}

