% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pesi_distributions}
\alias{pesi_distributions}
\title{Get PESI distributions from GUIDs}
\usage{
pesi_distributions(guid = NULL, opts = NULL, iface = NULL, ...)
}
\arguments{
\item{guid}{One or more PESI GUIDs}

\item{opts}{Options passed on to \code{SSOAP::.SOAP} for curl debugging.}

\item{iface}{PESI SOAP interface. By default uses the cached version in taxize.}

\item{...}{Further args passed on to SSOAP}
}
\description{
Get PESI distributions from GUIDs
}
\examples{
\dontrun{
pesi_distributions(guid='A0433E13-D7B5-49F2-86BA-A1777364C559')
pesi_distributions(guid=c('A0433E13-D7B5-49F2-86BA-A1777364C559',
                   '66374558-8F9A-4833-AEC8-490DAAC76024'))
}
}

