% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pesi_gen_iface}
\alias{pesi_gen_iface}
\title{Generate PESI function interface.}
\usage{
pesi_gen_iface(wsdl_url = "http://www.eu-nomen.eu/portal/soap.php?wsdl=1",
  ...)
}
\arguments{
\item{wsdl_url}{URL for the PESI SOAP WSDL file}

\item{...}{Further args passed on to \code{genSOAPClientInterface}}
}
\value{
Returns invisibly a S4 object holding all functions to interact with WORMS.
}
\description{
Generate PESI function interface.
}
\examples{
\dontrun{
# You can generate your own interface to PESI functions
out <- pesi_gen_iface()
out

# Then access them directly
out@functions$getPESIRecords(scientificname='Salmo', like=FALSE)

# Or pass to the taxize functions in the iface parameter
head(pesi_records(scientific = "Salmo", iface = out))
}
}

