% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pesi_guid}
\alias{pesi_guid}
\title{Get PESI GUID from scientific names}
\usage{
pesi_guid(scientific = NULL, opts = NULL, iface = NULL, ...)
}
\arguments{
\item{scientific}{One of more scientific names}

\item{opts}{Options passed on to \code{SSOAP::.SOAP} for curl debugging.}

\item{iface}{PESI SOAP interface. By default uses the cached version in taxize.}

\item{...}{Further args passed on to SSOAP}
}
\description{
Get PESI GUID from scientific names
}
\examples{
\dontrun{
pesi_guid(scientific='Ternatea vulgaris')
pesi_guid(scientific='Holcus lanatus')
pesi_guid(scientific=c('Ternatea vulgaris','Carcharhinus altimus','Carcharhinus galapagensis'))
}
}

