% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pesi_name_scientific}
\alias{pesi_name_scientific}
\title{Get PESI scientific names from GUIDs}
\usage{
pesi_name_scientific(guid = NULL, opts = NULL, iface = NULL, ...)
}
\arguments{
\item{guid}{One or more PESI GUIDs}

\item{opts}{Options passed on to \code{SSOAP::.SOAP} for curl debugging.}

\item{iface}{PESI SOAP interface. By default uses the cached version in taxize.}

\item{...}{Further args passed on to SSOAP}
}
\description{
Get PESI scientific names from GUIDs
}
\examples{
\dontrun{
pesi_name_scientific(guid='67C3CC9C-624A-40C5-B63A-AB880E0300D1')
pesi_name_scientific(guid=c('urn:lsid:marinespecies.org:taxname:105787','
                 urn:lsid:marinespecies.org:taxname:105790'))
}
}

