% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pesi_records}
\alias{pesi_records}
\title{Get PESI records from scientific names, common names, or GUIDs}
\usage{
pesi_records(scientific = NULL, common = NULL, guid = NULL, like = NULL,
  opts = NULL, iface = NULL, ...)
}
\arguments{
\item{scientific}{One of more scientific names}

\item{common}{One of more scientific names}

\item{guid}{One or more PESI GUIDs}

\item{like}{Argument...}

\item{opts}{Options passed on to \code{SSOAP::.SOAP} for curl debugging.}

\item{iface}{PESI SOAP interface. By default uses the cached version in taxize.}

\item{...}{Further args passed on to SSOAP}
}
\description{
Get PESI records from scientific names, common names, or GUIDs
}
\examples{
\dontrun{
pesi_records(scientific='Ternatea vulgaris')
pesi_records(common='arctic tern')
pesi_records(common='great white shark')
pesi_records(guid='67C3CC9C-624A-40C5-B63A-AB880E0300D1')
pesi_records(guid=c('urn:lsid:marinespecies.org:taxname:105787',
                   'urn:lsid:marinespecies.org:taxname:105790'))
}
}

