% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pesi_search}
\alias{pesi_search}
\title{Search for PESI scientific names and associated metadata.}
\usage{
pesi_search(scientific = NULL, opts = NULL, iface = NULL, ...)
}
\arguments{
\item{scientific}{One or more scientific names}

\item{opts}{Options passed on to \code{SSOAP::.SOAP} for curl debugging.}

\item{iface}{PESI SOAP interface. By default uses the cached version in taxize.}

\item{...}{Further args passed on to SSOAP}
}
\description{
Search for PESI scientific names and associated metadata.
}
\examples{
\dontrun{
pesi_search(scientific='Ternatea vulgaris')
pesi_search(scientific=c('Ternatea vulgaris','Carcharhinus altimus','Carcharhinus galapagensis'))
}
}

