% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomatic_format.R
\name{phylomatic_format}
\alias{phylomatic_format}
\title{Get family names to make Phylomatic input object, and output input string
to Phylomatic for use in the function phylomatic_tree.}
\usage{
phylomatic_format(...)
}
\arguments{
\item{...}{Parameters, ignored}
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\keyword{internal}
