% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{bold_ping}
\alias{col_ping}
\alias{eol_ping}
\alias{gbif_ping}
\alias{itis_ping}
\alias{nbn_ping}
\alias{ncbi_ping}
\alias{ping}
\alias{tropicos_ping}
\alias{ubio_ping}
\title{Ping an API used in taxize to see if it's working.}
\usage{
col_ping(what = "status", ...)

eol_ping(what = "status", ...)

itis_ping(what = "status", ...)

ncbi_ping(what = "status", ...)

tropicos_ping(what = "status", ...)

nbn_ping(what = "status", ...)

gbif_ping(what = "status", ...)

ubio_ping(what = "status", ...)

bold_ping(what = "status", ...)
}
\arguments{
\item{what}{(character) One of status (default), content, or an HTTP status code. If status,
we just check that the HTTP status code is 200, or similar signifying the service is up.
If content, we do a simple, quick check to determine if returned content matches what's
expected. If an HTTP status code, it must match an appropriate code.
See \code{\link{http_codes}}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A logical, TRUE or FALSE
}
\description{
Ping an API used in taxize to see if it's working.
}
\details{
For ITIS, see \code{\link{getdescription}}, which provides number of
scientific and common names in a character string.
}
\examples{
\dontrun{
col_ping()
col_ping("content")
col_ping(200)
col_ping("200")
col_ping(204)

itis_ping()
eol_ping()
ncbi_ping()
tropicos_ping()
nbn_ping()

gbif_ping()
gbif_ping(200)
ubio_ping()

bold_ping()
bold_ping(200)
bold_ping("content")
}
}

