% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{plantGenusNames}
\alias{plantGenusNames}
\title{Vector of plant genus names from ThePlantList}
\format{
A vector of length 793
}
\source{
http://www.theplantlist.org
}
\description{
These names are from http://www.theplantlist.org, and are a
randomly chosen subset of genera names for the purpose of having some
names to play with for examples in this package.
}
\keyword{data}
