% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantminer.R
\name{plantminer-defunct}
\alias{plantminer-defunct}
\alias{plantminer}
\title{Search for taxonomy data from Plantminer.com}
\usage{
plantminer(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Search for taxonomy data from Plantminer.com
}
