% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{plantNames}
\alias{plantNames}
\title{Vector of plant species (genus - specific epithet) names from ThePlantList}
\format{
A vector of length 1182
}
\source{
http://www.theplantlist.org
}
\description{
These names are from http://www.theplantlist.org, and are a
randomly chosen subset of names of the form genus/specific epithet
for the purpose of having some names to play with for examples in
this package.
}
\keyword{data}
