% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_search.R
\name{pow_synonyms}
\alias{pow_synonyms}
\title{Lookup synonyms in Kew's Plants of the World}
\usage{
pow_synonyms(id, ...)
}
\arguments{
\item{id}{(character) taxon id. required}

\item{...}{Further args passed on to \code{\link[=pow_lookup]{pow_lookup()}}}
}
\description{
Lookup synonyms in Kew's Plants of the World
}
\examples{
\dontrun{
pow_synonyms(id = 'urn:lsid:ipni.org:names:320035-2')
pow_synonyms(id = 'urn:lsid:ipni.org:names:358881-1')
pow_synonyms(id = 'urn:lsid:ipni.org:names:359855-1')
}
}
\seealso{
Other pow: 
\code{\link{get_pow}()},
\code{\link{pow_lookup}()},
\code{\link{pow_search}()}
}
\concept{pow}
