% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{rank_ref_zoo}
\alias{rank_ref_zoo}
\title{Lookup-table for IDs of taxonomic ranks (WoRMS)}
\description{
Same as \code{rank_ref} but specifically for WoRMS, where section/subsection
ranks are put between family/order rather than between species/genus.
}
\keyword{data}
