% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scamit_search}
\alias{scamit_search}
\title{Search SCAMIT data for taxonomic IDs}
\usage{
scamit_search(name, rank = "genus", ...)
}
\arguments{
\item{name}{The string to search for.}

\item{rank}{Taxonomic rank to search against. One of phylum, subphylum, class, subclass,
infraclass, superorder, order, suborder, infraorder, superfamily, family, subfamily,
tribe, genus, subgenus, or species}

\item{...}{Further args passed on to \code{agrep()}. See \code{?agrep}}
}
\description{
SCAMIT is the Southern California Association of Marine Invertebrate Taxonomists. Find out more
at \url{http://www.scamit.org/}
}
\note{
Note that we set \code{ignore.case=TRUE} by default.
}
\examples{
# Compare
scamit_search(name="Cliona", rank="genus")
scamit_search(name="Cliona", rank="genus", max.distance = 0)
scamit_search(name="Silicea", rank="phylum")
}

