% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{searchbycommonname}
\alias{searchbycommonname}
\title{Search for tsn by common name}
\usage{
searchbycommonname(x, backend = NULL, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for tsn by common name
}
\examples{
\dontrun{
searchbycommonname("american bullfrog", config=timeout(3))
searchbycommonname("ferret-badger", config=timeout(3))
searchbycommonname("polar bear", "api")

searchbycommonname("american bullfrog", "local")
searchbycommonname("ferret-badger", "local")
searchbycommonname("polar bear", "local")
}
}
\keyword{internal}

