% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{searchbycommonnamebeginswith}
\alias{searchbycommonnamebeginswith}
\title{Search for tsn by common name beginning with}
\usage{
searchbycommonnamebeginswith(x, backend = NULL, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or localsql, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for tsn by common name beginning with
}
\examples{
\dontrun{
searchbycommonnamebeginswith("inch", "api")
searchbycommonnamebeginswith("inch", "local")
}
}
\keyword{internal}

