% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{searchbyscientificname}
\alias{searchbyscientificname}
\title{Search by scientific name}
\usage{
searchbyscientificname(x, backend = NULL, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search by scientific name
}
\examples{
\dontrun{
searchbyscientificname("Tardigrada", "api")
searchbyscientificname("Tardigrada", "local")
searchbyscientificname("Quercus douglasii", "api")
searchbyscientificname("Quercus douglasii", "local")
}
}
\keyword{internal}

