% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{searchforanymatch}
\alias{searchforanymatch}
\title{Search for any match}
\usage{
searchforanymatch(x, backend = NULL, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{backend}{Defaults to NULL, deferring to options set by \code{\link{backend_set}}.
Alternatively, you can pass in one of api or local, which will only override the
current function call.}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for any match
}
\examples{
\dontrun{
searchforanymatch(202385, "api")
searchforanymatch(202385, "local")
searchforanymatch("dolphin", "api")
searchforanymatch("dolphin", "local")
}
}
\keyword{internal}

