% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchforanymatchpaged}
\alias{searchforanymatchpaged}
\title{Search for any matched page}
\usage{
searchforanymatchpaged(srchkey = NA, pagesize = NA, pagenum = NA,
  ascend = NA, curlopts = list(), curl = getCurlHandle())
}
\arguments{
\item{pagesize}{An integer containing the page size (numeric)}

\item{pagenum}{An integer containing the page number (numeric)}

\item{ascend}{A boolean containing true for ascending sort order or false
for descending (logical)}
}
\description{
Search for any matched page
}
\examples{
\donttest{
searchforanymatchpaged(srchkey=202385, pagesize=100, pagenum=1, ascend=FALSE)
searchforanymatchpaged(srchkey="Zy", pagesize=100, pagenum=1, ascend=FALSE)
}
}
\keyword{internal}

