% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_rank2.R
\name{tax_rank2}
\alias{tax_rank2}
\alias{tax_rank2.default}
\alias{tax_rank2.default}
\alias{tax_rank2.character}
\alias{tax_rank2.numeric}
\title{get rank names}
\usage{
tax_rank2(...)

\method{tax_rank2}{default}(x, db = NULL, ...)

\method{tax_rank2}{default}(x, db = NULL, ...)

\method{tax_rank2}{character}(x, db = NULL, ...)

\method{tax_rank2}{numeric}(x, db = NULL, ...)
}
\arguments{
\item{...}{Additional arguments to \code{\link{classification}}}

\item{x}{(character) Vector of one or more taxon names (character) or
IDs (character or numeric) to query.}

\item{db}{(character) database to query. either \code{ncbi}, \code{itis},
\code{eol}, \code{col}, \code{tropicos}, \code{gbif}, \code{nbn},
\code{worms}, \code{natserv}, \code{bold}. Note that each taxonomic data
source has their own identifiers, so that if you provide the wrong
\code{db} value for the identifier you could get a result, but it will
likely be wrong (not what you were expecting).}
}
\value{
A named list of character vectors with ranks (all lower-cased)
}
\description{
get rank names
}
\examples{
\dontrun{
tax_rank2(x = "Helianthus annuus", db = "itis")
tax_rank2(get_tsn("Helianthus annuus"))
tax_rank2(c("Helianthus", "Pinus", "Poa"), db = "itis")

tax_rank2(get_boldid("Helianthus annuus"))
tax_rank2("421377", db = "bold")
tax_rank2(421377, db = "bold")

tax_rank2(c("Helianthus annuus", "Puma", "Homo sapiens"), db = 'ncbi')
tax_rank2(c("Helianthus annuus", "Quercus", "Fabaceae"), db = 'tropicos')

tax_rank2(names_list("species"), db = 'gbif')
tax_rank2(names_list("family"), db = 'gbif')

tax_rank2(c("Platanista gangetica", "Lichenopora neapolitana"), db = "worms")
}
}
