% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\name{taxize-deprecated}
\alias{taxize-deprecated}
\alias{defunct}
\title{Deprecated functions in taxize}
\description{
The following functions are now deprecated:
}
\details{
\itemize{
\item \code{\link[=get_tsn]{get_tsn()}}: use \code{\link[=get_itis]{get_itis()}}
\item \code{\link[=get_uid]{get_uid()}}: use \code{\link[=get_ncbi]{get_ncbi()}}
}
}
