% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-params.R
\name{taxize-params}
\alias{taxize-params}
\title{taxize parameters}
\description{
Information on standardized parameters across the package
}
\section{Standardized parameters}{

\itemize{
\item \code{sci}: scientific name
\item \code{com}: common name
\item \code{id}: name identifier
\item \code{sci_com}: scientific name or common name
\item \code{sci_id}: scientific name or name identifier
}

We were going to standardize parameter names for cases in which
a parameter accepts either of three options: scientific name,
common name, or name identifier. However, there was no clear
parameter name we could use for this case, so we've left
parameter names as they are for the two cases (\code{\link[=get_ids]{get_ids()}} and
\code{\link[=vascan_search]{vascan_search()}})
}

