% Generated by roxygen2 (4.0.1): do not edit by hand
\name{taxize_capwords}
\alias{taxize_capwords}
\title{Capitalize the first letter of a character string.}
\usage{
taxize_capwords(s, strict = FALSE, onlyfirst = FALSE)
}
\arguments{
\item{s}{A character string}

\item{strict}{Should the algorithm be strict about capitalizing.
Defaults to FALSE.}

\item{onlyfirst}{Capitalize only first word, lowercase all others. Useful
for taxonomic names.}
}
\description{
Capitalize the first letter of a character string.
}
\examples{
taxize_capwords(c("using AIC for model selection"))
taxize_capwords(c("using AIC for model selection"), strict=TRUE)
}

