% Generated by roxygen2 (4.0.1): do not edit by hand
\name{taxize_cite}
\alias{taxize_cite}
\title{Get citations and licenses for data sources used in taxize}
\usage{
taxize_cite(fxn = "itis", what = "citation")
}
\arguments{
\item{fxn}{}

\item{what}{One of citation (default), license, or both.}
}
\description{
Get citations and licenses for data sources used in taxize
}
\examples{
taxize_cite(fxn='tnrs')
taxize_cite(fxn='eol')

# Get the taxize citation
taxize_cite(fxn='taxize')

# Get license information
taxize_cite(fxn='taxize', "license")
taxize_cite(fxn='tnrs', "license")
}

