% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressor.R
\name{taxize_options}
\alias{taxize_options}
\title{taxon state options}
\usage{
taxize_options(taxon_state_messages = NULL, quiet = FALSE)
}
\arguments{
\item{taxon_state_messages}{(logical) suppress messages? default: \code{NULL}
(same as setting \code{FALSE}). Set to \code{TRUE} to suppress messages, and \code{FALSE}
to not suppress messages}

\item{quiet}{(logical) quiet informational output from this function.
default: \code{TRUE}}
}
\description{
taxon state options
}
\examples{
taxize_options()
taxize_options(FALSE)
taxize_options(TRUE)
}
\keyword{internal}
