% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_state.R
\name{taxon_last}
\alias{taxon_last}
\title{Get the last taxon state object}
\usage{
taxon_last()
}
\value{
an object of class `taxon_state`, the last one used, else `NULL`
if none found
}
\description{
Get the last taxon state object
}
\keyword{internal}
