% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_state.R
\docType{data}
\name{taxon_state}
\alias{taxon_state}
\title{Keep track of queries in `get_*` functions}
\description{
This object lives inside each `get_*` function call, maintaining
results as they are accumulated.
}
\details{
**Methods**
  \describe{
    \item{`add(query, result)`}{
      add a query with it's result
    }
    \item{`get(query)`}{
      get a result by query
    }
    \item{`remove(query)`}{
      remove a query
    }
    \item{`purge()`}{
      remove all records
    }
    \item{`count` (active binding)}{
      count number of records
    }
    \item{`exit` (active binding)}{
      record date/time function exited
    }
  }
}
\examples{
ts <- taxon_state$new()
taxon_last()
ts
res <- list(
  id = 123456,
  att = "found",
  multiple = FALSE,
  direct = FALSE,
  class = "tsn"
)
ts$add(query = "Quercus robur", result = res)
ts
ts$get(query = "Quercus robur")
ts$count
ts$remove(query = "Quercus robur")
ts
ts$count

res2 <- list(
  id = 3430834535,
  att = "found",
  multiple = FALSE,
  direct = FALSE,
  class = "gbifid"
)
ts$add(query = "Poa annua", result = res2)
res3 <- list(
  id = 1223424,
  att = "found",
  multiple = FALSE,
  direct = FALSE,
  class = "uid"
)
ts$add(query = "Puma concolor", result = res3)
ts
ts$count
ts$get("Puma concolor")
ts$get()

# cleanup
ts$purge()
ts$count
}
\keyword{internal}
