% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnrs_sources.r
\name{tnrs_sources}
\alias{tnrs_sources}
\title{TNRS sources}
\usage{
tnrs_sources(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Defunct - service was down far too much to be reliable
}
\keyword{internal}
