% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnrs_sources.r
\name{tnrs_sources}
\alias{tnrs_sources}
\title{TNRS sources}
\usage{
tnrs_sources(source = NULL, ...)
}
\arguments{
\item{source}{The source to get information on, one of "iPlant_TNRS",
"NCBI", or "MSW3".}

\item{...}{Curl options to pass in \code{\link[httr]{GET}}}
}
\value{
Sources for the TNRS API in a vector or list
}
\description{
Get sources for the Phylotastic Taxonomic Name Resolution Service
}
\examples{
\dontrun{
# All
tnrs_sources()

# A specific source
tnrs_sources(source="NCBI")
}
}

