% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol_resolve.R
\name{tol_resolve}
\alias{tol_resolve}
\title{Resolve names using Open Tree of Life resolver}
\usage{
tol_resolve(names = NULL, context_name = NULL,
  do_approximate_matching = TRUE, ids = NULL, include_deprecated = FALSE,
  include_dubious = FALSE, ...)
}
\arguments{
\item{names}{taxon names to be queried (character vector)}

\item{context_name}{name of the taxonomic context to be searched
(length-one character vector). Must match (case sensitive) one
of the values returned by \code{\link[rotl]{tnrs_contexts}}.}

\item{do_approximate_matching}{A logical indicating whether or not
to perform approximate string (a.k.a. \dQuote{fuzzy})
matching. Using \code{FALSE} will greatly improve
speed. Default, however, is \code{TRUE}.}

\item{ids}{An array of ids to use for identifying names. These
will be assigned to each name in the names array. If ids is
provided, then ids and names must be identical in length.}

\item{include_deprecated}{A boolean indicating whether or not to
include deprecated taxa in the search.}

\item{include_dubious}{Whether to include so-called 'dubious'
taxa--those which are not accepted by OTT.}

\item{...}{Curl options passed on to \code{\link[httr]{POST}} within
\code{\link[rotl]{tnrs_match_names}}}
}
\value{
A data frame summarizing the results of the query. The
    original query output is appended as an attribute to the
    returned object (and can be obtained using \code{attr(object,
    "original_response")}).
}
\description{
Resolve names using Open Tree of Life resolver
}
\examples{
\dontrun{
tol_resolve(names=c("echinodermata", "xenacoelomorpha", "chordata", "hemichordata"))
tol_resolve(c("Hyla", "Salmo", "Diadema", "Nautilus"))
tol_resolve(c("Hyla", "Salmo", "Diadema", "Nautilus"), context_name = "Animals")

turducken_spp <- c("Meleagris gallopavo", "Anas platyrhynchos", "Gallus gallus")
tnrs_match_names(turducken_spp, context="Animals")
}
}
\author{
Francois Michonneau \email{francois.michonneau@gmail.com}
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\references{
\url{https://github.com/OpenTreeOfLife/opentree/wiki/Open-Tree-of-Life-APIs#match_names}
}
\seealso{
\code{\link{gnr_resolve}}, \code{\link{tnrs}}
}
\keyword{names}
\keyword{resolve}
\keyword{taxonomy}

