% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_accnames.R
\name{tp_acceptednames}
\alias{tp_acceptednames}
\title{Return all accepted names for a taxon name with a given id.}
\usage{
tp_acceptednames(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Function name changed to tp_accnames.
}
\keyword{internal}
