% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tp_accnames}
\alias{tp_accnames}
\title{Return all accepted names for a taxon name with a given id.}
\usage{
tp_accnames(id, key = NULL, callopts = list())
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; loads from .Rprofile.}

\item{callopts}{Further args passed on to httr::GET}
}
\value{
List or dataframe.
}
\description{
Return all accepted names for a taxon name with a given id.
}
\examples{
\dontrun{
tp_accnames(id = 25503923)
tp_accnames(id = 25538750)

# No accepted names found
tp_accnames(id = 25509881)
}
}

