% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_classification.R
\name{tp_classification}
\alias{tp_classification}
\title{Return all synonyms for a taxon name with a given id.}
\usage{
tp_classification(...)
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\keyword{internal}

