% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tp_namedistributions}
\alias{tp_namedistributions}
\title{Return all distribution records for for a taxon name with a given id.}
\usage{
tp_namedistributions(id, key = NULL, callopts = list())
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; loads from .Rprofile. Or you can passin your
key in this arg.}

\item{callopts}{Further args passed on to httr::GET}
}
\description{
Function name changed to tp_dist.
}
\keyword{internal}

