% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_refs.R
\name{tp_namereferences}
\alias{tp_namereferences}
\title{Return all reference records for for a taxon name with a given id.}
\usage{
tp_namereferences(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Function name changed to tp_refs.
}
\keyword{internal}
