% Generated by roxygen2 (4.0.0): do not edit by hand
\name{try_with_timeout}
\alias{try_with_timeout}
\title{Function to allow user to set timeout on a function call.}
\usage{
try_with_timeout(test_fn, tlimit = 30, defaultvalue = "TimedOut")
}
\arguments{
\item{test_fn}{The function call to test}

\item{tlimit}{= 120 A timeout in seconds}
}
\value{
If in a loop, passes on to next thing, or if not, then gives back message
}
\description{
This is separate from the curl timeout.ms parameter
}
\examples{
\dontrun{
library("taxize")
try_with_timeout(eol_search('Salix'))
try_with_timeout(eol_search('Salix'), tlimit=10)
}
}

