% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_taxon.R
\name{txid}
\alias{txid}
\alias{as.data.frame.txid}
\title{txid}
\usage{
\method{as.data.frame}{txid}(x, ...)
}
\arguments{
\item{x}{an object of class \code{txid}}

\item{...}{ignored}
}
\value{
a tibble
}
\description{
txid is the class all \verb{get_*()} functions return. It is extended from
\code{\link[vctrs:new_rcrd]{vctrs::new_rcrd()}} and is adapted from \code{\link[taxa2:taxon]{taxa2::taxon()}} with modificaitons
specifically for taxize.
}
\details{
Additional classes attached include \code{taxa_taxon} (mostly to differentiate
it from objects created from \code{\link[taxa2:taxon]{taxa2::taxon()}}), vctrs classes
\code{vctrs_rcrd} and \code{vctrs_vctr} (because rcrd is extended from vctr),
and a class for which data source the data is from (e.g., \code{gbif}).

See \link{id-accessors} for functions for accessing parts of txid objects
}
\seealso{
many taxize functions have S3 methods for \code{txid}, e.g.,
\link{children}
}
