% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubio_classification.R
\name{ubio_classification}
\alias{ubio_classification}
\title{uBio classification}
\usage{
ubio_classification(...)
}
\arguments{
\item{...}{Parameters, ignored}
}
\description{
THIS FUNCTION IS DEFUNCT.
}
