% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ubio_classification_search.R
\name{ubio_classification_search}
\alias{ubio_classification_search}
\title{This function will return ClassificationBankIDs (hierarchiesIDs) that refer to the
given NamebankID}
\usage{
ubio_classification_search(...)
}
\arguments{
\item{...}{Parameters, ignored}
}
\description{
THIS FUNCTION IS DEFUNCT.
}

