% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ubio_search}
\alias{ubio_search}
\title{This function will return NameBankIDs that match given search terms}
\usage{
ubio_search(searchName = NULL, searchAuth = NULL, searchYear = NULL,
  order = NULL, sci = NULL, vern = NULL, keyCode = NULL,
  callopts = list())
}
\arguments{
\item{searchName}{(string) - term to search within name string}

\item{searchAuth}{(string) - term to search within name authorship}

\item{searchYear}{(string) - term to search within name year}

\item{order}{(string) - (name or namebankID) field by which the results will
be sorted (default is namebankID)}

\item{sci}{(int) - (sci, vern, or all) type of results to be returned
(default is all)}

\item{vern}{(int) - (limit 1000) maximum number of results to be returned
(default is 1000)}

\item{keyCode}{Your uBio API key; loads from .Rprofile. If you don't have
one, obtain one at http://www.ubio.org/index.php?pagename=form.}

\item{callopts}{Parameters passed on to httr::GET call.}
}
\value{
A data.frame.
}
\description{
This function will return NameBankIDs that match given search terms
}
\examples{
\dontrun{
ubio_search(searchName = 'elephant', sci = 1, vern = 0)
ubio_search(searchName = 'Astragalus aduncus', sci = 1, vern = 0)
}
}

