% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{wiki_data}
\alias{get_wiki_id}
\alias{wiki_data}
\title{Wikipedia taxonomy data}
\usage{
wiki_data(x, property = NULL, ...)

get_wiki_id(x, language = "en", limit = 10, ...)
}
\arguments{
\item{x}{(character) a taxonomic name}

\item{property}{(character) a property id, e.g., P486}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}

\item{language}{(character) two letter language code}

\item{limit}{(integer) records to return. Default: 10}
}
\description{
Wikipedia taxonomy data
}
\examples{
\dontrun{
wiki_data("Poa annua")
wiki_data("Mimulus alsinoides")
wiki_data("Mimulus foliatus")
wiki_data("Mimulus foliatus", property = "P846")
wiki_data("Mimulus foliatus", property = c("P846", "P815"))

wiki_data(get_wiki_id("Mimulus foliatus"))
}
}

