% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms}
\alias{worms}
\title{Search World Register of Marine Species (WoRMS)}
\description{
WORMS has a SOAP API. We store the machine generated API specification in the package as the
object \code{worms_iface}. However, you can update the spec if you want using
\code{worms_gen_iface}, then pass the output of that fxn to the \code{iface} parameter of
\code{worms_*} functions.
}
\details{
The following functions are available to interact with WoRMS:

\itemize{
 \item worms_gen_iface Generate new WoRMS SOAP API interface.
 \item worms_children Get immediate children from a WoRMS ID.
 \item worms_common  Get common (vernacular) names from a WoRMS ID.
 \item worms_extid  Get external IDs from a WoRMS ID, E.g., get an NCBI identifier.
 \item worms_hierarchy  Get complete taxonomic hierarchy from a WoRMS ID.
 \item worms_name  Get scientific name from a WoRMS ID.
 \item worms_records Get WoRMS records from a WoRMS ID, an external ID, a scientific name, a
 common name, or a start- or end-date.
 \item worms_sources Get sources (references) for taxa from a WoRMS ID.
 \item worms_synonyms Get scientific name synonyms from a WoRMS ID.
}
}
\references{
\url{http://www.marinespecies.org/}
}

