% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_children}
\alias{worms_children}
\title{Children search}
\usage{
worms_children(ids = NULL, offset = NULL, marine_only = 1, opts = NULL,
  update_iface = FALSE)
}
\description{
Children search
}
\examples{
\dontrun{
worms_children(ids=106135)
}
}

