% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_common}
\alias{worms_common}
\title{Common names from ID}
\usage{
worms_common(ids = NULL, opts = NULL, update_iface = FALSE, ...)
}
\arguments{
\item{id}{(numeric) One or more worms AphidID for a taxon.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function
which actually invokes the SOAP method. These options control aspects of the HTTP request,
including debugging information that is displayed on the console,
e.g. .opts = list(verbose = TRUE)}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}
}
\description{
Common names from ID
}
\examples{
\dontrun{
worms_common(ids=1080)
worms_common(ids=22388)
worms_common(ids=123080)
worms_common(ids=160281)
}
}

