% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_gen_iface}
\alias{worms_gen_iface}
\title{Generate WORMS interface.}
\usage{

  worms_gen_iface(wsdl_url = "http://www.marinespecies.org/aphia.php?p=soap&wsdl=1",
  ...)
}
\arguments{
\item{wsdl_url}{URL for the WORMS SOAP WSDL file}

\item{...}{Further args passed on to \code{genSOAPClientInterface}}
}
\value{
Returns invisibly a S4 object holding all functions to interact with WORMS.
}
\description{
Generate WORMS interface.
}
\examples{
\dontrun{
out <- worms_gen_iface()
out
worms_records(scientific='Salmo', iface=out)
}
}

