% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{worms_iface}
\alias{worms_iface}
\title{Stored WoRMS SOAP API interface.}
\description{
An S4 class generated from \code{worms_gen_iface}. You can navigate to each function by doing,
for example \code{worms_gen_iface()@functions$getAphiaID}. Note that there will be some warnings
thrown, but it should still work. It is not easy to suppress those warnings as they are printed
with \code{cat}. You can dig into each of the below functions, explore
the code for each function, explore it's parameters, etc. The functions:
}
\details{
\itemize{
 \item getAphiaID
 \item getAphiaRecords
 \item getAphiaNameByID
 \item getAphiaRecordByID
 \item getAphiaRecordByExtID
 \item getExtIDbyAphiaID
 \item getAphiaRecordsByNames
 \item getAphiaRecordsByVernacular
 \item getAphiaRecordsByDate
 \item getAphiaClassificationByID
 \item getSourcesByAphiaID
 \item getAphiaSynonymsByID
 \item getAphiaVernacularsByID
 \item getAphiaChildrenByID
 \item matchAphiaRecordsByNames
}

The above functions are used internally to make more user friendly function interfaces
to WORMS data. See \code{worms_*} for all worms functions.
}
\keyword{data}

