% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_name}
\alias{worms_name}
\title{Get name from a WORMS id}
\usage{
worms_name(ids = NULL, opts = NULL, iface = NULL, ...)
}
\arguments{
\item{ids}{(numeric) One or more WoRMS AphidID's for a taxon.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function
which actually invokes the SOAP method. These options control aspects of the HTTP request,
including debugging information that is displayed on the console,
e.g. .opts = list(verbose = TRUE)}

\item{iface}{Interface to WoRMS SOAP API methods. By default we use a previously created object.
If you want to create a new one, use \code{worms_gen_iface}, assign the output to an object,
then pass it into any \code{worms_*} function. in the \code{iface} parameter.}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}
}
\description{
Get name from a WORMS id
}
\examples{
\dontrun{
worms_name(ids=1080)
worms_name(ids=c(1080,22388,160281,123080,22388))
worms_name(ids=c(1080,22388,160281,123080,22388,106135,159283))
}
}

