% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_records}
\alias{worms_records}
\title{Get records by ID, scientific name, common name, date, worms id, or external id.}
\usage{
worms_records(scientific = NULL, common = NULL, id = NULL, extid = NULL,
  like = NULL, type = NULL, marine_only = 1, offset = NULL,
  startdate = NULL, enddate = NULL, opts = NULL, update_iface = FALSE,
  ...)
}
\arguments{
\item{id}{(numeric) One or more worms AphidID for a taxon.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function
which actually invokes the SOAP method. These options control aspects of the HTTP request,
including debugging information that is displayed on the console,
e.g. .opts = list(verbose = TRUE)}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}
}
\description{
Get records by ID, scientific name, common name, date, worms id, or external id.
}
\examples{
\dontrun{
worms_records(scientific='Salmo')
worms_records(scientific='Liopsetta glacialis')
worms_records(common='salmon')
worms_records(startdate='2014-06-01T00:00:00', enddate='2014-06-02T00:00:00')
worms_records(id=1080)
worms_records(extid=6830, type='ncbi')
}
}

