% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_search_name}
\alias{worms_search_name}
\title{Search World Register of Marine Species (WoRMS)}
\usage{
worms_search_name(names = NULL, by = "getAphiaID", update_iface = FALSE)
}
\arguments{
\item{names}{asdadsf}

\item{by}{asdfda}

\item{update_iface}{ssdafdf}
}
\value{
A data.frame.
}
\description{
Search World Register of Marine Species (WoRMS)
}
\details{
You must provide one of name or id. The other parameters (format
		and start) are optional. \code{getAphiaRecordByExtID} is not available through this fxn.
  	Search by AphiaID, scientific name, vernacular name, or date.
}
\examples{
\dontrun{
# A basic example
worms_search(names="Apis")
worms_search(names='', by='getAphiaRecordsByNames')
}
}
\references{
\url{http://www.marinespecies.org/}
}

