% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{worrms_ranks}
\alias{worrms_ranks}
\title{WORMS ranks}
\format{A data frame with 494 rows and 3 variables:
\itemize{
\item \code{id}: rank id
\item \code{rank}: rank name
}}
\description{
Created using \code{worrms::wm_ranks_id(-1)} on 2020-02-11.
}
\details{
Present in taxize in the case where WORMS does not
return rank names - with this dataset we can fill
in rank information as long as rank ids are returned
}
\keyword{data}
