% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apg.R
\name{apg}
\alias{apg}
\alias{apgFamilies}
\alias{apgOrders}
\title{Get APG names}
\usage{
apgOrders(...)

apgFamilies(...)
}
\arguments{
\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
FIXME: add function for genera
}
\details{
Generic names and their replacements from the Angiosperm Phylogeny
Group III system of flowering plant classification.
}
\examples{
\dontrun{
head(apgOrders())
head(apgFamilies())
}
}
\references{
\url{http://www.mobot.org/MOBOT/research/APweb/}
}

