% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{apg_families}
\alias{apg_families}
\title{MOBOT family names}
\format{
A data frame with 1705 rows and 6 variables:
\itemize{
\item \code{family}: family name
\item \code{synonym}: if \code{accepted=FALSE}, this is the accepted name;
if \code{accepted=TRUE}, this is \code{NA}, and the name in \code{family} is accepted
\item \code{order}: order name for the family
\item \code{accepted}: logical, if name in \code{family} column is accepted or not
\item \code{original}: original data record from APG website, mapping
name in \code{family} column to a new name, if there is one
\item \code{accepted_name}: accepted name. accepted names, combining those that
are accepted from \code{family} column, with the new name from \code{synonym}
if applicable
}
}
\source{
\url{http://www.mobot.org/MOBOT/research/APweb/}
}
\description{
Family names and their replacements from the Angiosperm Phylogeny
Website system of flowering plant classification.
}
\details{
This dataset is from Version 14, incorporated on 2020-06-03,
generated using \code{\link[=apgFamilies]{apgFamilies()}}
}
\keyword{data}
