\name{classification}
\alias{classification}
\title{Get taxonomic hierarchy for a given taxon.}
\usage{
  classification(x, ...)
}
\arguments{
  \item{x}{quoted tsn number (taxonomic serial number)}
}
\value{
  Classification of the taxon in a data.frame.
}
\description{
  Get taxonomic hierarchy for a given taxon.
}
\examples{
\dontrun{
classification(685566)
}
}

