% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_downstream.R
\name{col_downstream}
\alias{col_downstream}
\title{Use Catalogue of Life to get downstream taxa to a given taxonomic level.}
\usage{
col_downstream(name = NULL, id = NULL, downto, format = NULL,
  start = NULL, checklist = NULL, verbose = TRUE, intermediate = FALSE,
  ...)
}
\arguments{
\item{name}{The string to search for. Only exact matches found the name given
 will be returned, unless one or wildcards are included in the search
   string. An * (asterisk) character denotes a wildcard; a % (percentage)
character may also be used. The name must be at least 3 characters long,
not counting wildcard characters.}

\item{id}{The record ID of the specific record to return (only for scientific
names of species or infraspecific taxa)}

\item{downto}{The taxonomic level you want to go down to. See examples below.
The taxonomic level IS case sensitive, and you do have to spell it
correctly. See \code{data(rank_ref)} for spelling.}

\item{format}{The returned format (default = NULL). If NULL xml is used.
Currently only xml is supported.}

\item{start}{The first record to return (default = NULL). If NULL, the
results are returned from the first record (start=0). This is useful if
the total number of results is larger than the maximum number of results
returned by a single Web service query (currently the maximum number of
results returned by a single query is 500 for terse queries and 50 for
full queries).}

\item{checklist}{The year of the checklist to query, if you want a specific
year's checklist instead of the lastest as default (numeric).}

\item{verbose}{Print or suppress messages.}

\item{intermediate}{(logical) If TRUE, return a list of length two with target
taxon rank names, with additional list of data.frame's of intermediate
taxonomic groups. Default: FALSE}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A list of data.frame's.
}
\description{
Use Catalogue of Life to get downstream taxa to a given taxonomic level.
}
\details{
Provide only names instead of id's
}
\examples{
\dontrun{
# Some basic examples
col_downstream(name="Apis", downto="Species")
col_downstream(name="Bryophyta", downto="Family")

# get classes down from the kingdom Animalia
col_downstream(name="Animalia", downto="Class")
col_downstream(name="Animalia", downto="Class", intermediate=TRUE)

# An example that takes a bit longer
col_downstream(name=c("Plantae","Animalia"), downto="Class")

# Using a checklist from a specific year
col_downstream(name="Bryophyta", downto="Family", checklist=2009)

# By id
col_downstream(id=2346405, downto="Genus", checklist=2012)
}
}

