% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eol_dataobjects.R
\name{eol_dataobjects}
\alias{eol_dataobjects}
\title{Given the identifier for a data object, return all metadata about the object.}
\usage{
eol_dataobjects(id, returntype = "data.frame", usekey = TRUE, key = NULL,
  verbose = TRUE)
}
\arguments{
\item{id}{The EOL data object identifier (character)}

\item{returntype}{one of "list" of "data.frame" (character)}

\item{usekey}{use your API key or not (TRUE or FALSE)}

\item{key}{Your EOL API key; loads from .Rprofile.}

\item{verbose}{logical; If TRUE the actual taxon queried is printed on the
console.}
}
\value{
List or dataframe (default).
}
\description{
Given the identifier for a data object, return all metadata about the object.
}
\details{
It's possible to return JSON or XML with the EOL API. However,
		this function only returns JSON for now.
}
\examples{
\dontrun{
eol_dataobjects(id="d72801627bf4adf1a38d9c5f10cc767f")
eol_dataobjects(id="21929584")
}
}

