% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apni.R
\name{get_apni}
\alias{get_apni}
\alias{as.apni}
\alias{as.apni.apni}
\alias{as.apni.character}
\alias{as.apni.list}
\alias{as.apni.data.frame}
\alias{as.data.frame.apni}
\alias{get_apni_}
\title{Get Australian Plant Names Index identifier for a taxon}
\usage{
get_apni(
  x,
  accepted = FALSE,
  ask = TRUE,
  messages = TRUE,
  rows = NA,
  family_filter = NULL,
  rank_filter = NULL,
  ...
)

as.apni(x, check = TRUE)

\method{as.apni}{apni}(x, check = TRUE)

\method{as.apni}{character}(x, check = TRUE)

\method{as.apni}{list}(x, check = TRUE)

\method{as.apni}{data.frame}(x, check = TRUE)

\method{as.data.frame}{apni}(x, ...)

get_apni_(x, messages = TRUE, rows = NA, ...)
}
\arguments{
\item{x}{character; A vector of common or scientific names. Or, a
\code{taxon_state} object (see \link{taxon-state})}

\item{accepted}{logical; If TRUE, removes names that are not accepted
valid names by ITIS. Set to \code{FALSE} (default) to give back both
accepted and unaccepted names.}

\item{ask}{logical; should \code{get_apni} be run in interactive mode?
If TRUE and more than one pow is found for teh species, the user is
asked for input. If FALSE NA is returned for multiple matches.}

\item{messages}{logical; should progress be printed?}

\item{rows}{numeric; Any number from 1 to infinity. If the default NA,
all rows are considered. Note that this function still only gives back
a pow class object with one to many identifiers. See
\code{\link[=get_apni_]{get_apni_()}} to get back all, or a subset,
of the raw data that you are presented during the ask process.}

\item{family_filter}{(character) A division (aka phylum) name to filter
data after retrieved from NCBI. Optional. See \code{Filtering} below.}

\item{rank_filter}{(character) A taxonomic rank name to filter data after
retrieved from NCBI. See \code{\link[=rank_ref]{rank_ref()}} for possible options.
Though note that some data sources use atypical ranks, so inspect the data
itself for options. Optional. See \code{Filtering} below.}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}

\item{check}{logical; Check if ID matches any existing on the DB, only
used in \code{\link[=as.apni]{as.apni()}}}
}
\value{
A vector of taxonomic identifiers as an S3 class.
If a taxon is not found an \code{NA} is given. If more than one identifier
is found the function asks for user input if \code{ask = TRUE}, otherwise
returns \code{NA}. If \code{ask=FALSE} and \code{rows} does not equal
\code{NA}, then a data.frame is given back, but not of the uid class, which
you can't pass on to other functions as you normally can.

See \code{\link{get_id_details}} for further details including
attributes and exceptions
}
\description{
Get Australian Plant Names Index identifier for a taxon
}
\section{Filtering}{

The parameters \code{family_filter} an\code{rank_filter}er are not
used in the search to the data provider, but are used in filtering the data down to a
subset that is closer to the target you want.  For these two parameters,
you can use regex strings since we use \code{\link[=grep]{grep()}} internally to match.
Filtering narrows down to the set that matches your query, and removes the rest.
}

\examples{
\dontrun{
get_apni(x = "Acacia")
get_apni(c("Acacia","Poa"))

# Get back a subset
get_apni(x="Acacia", rows = 1)
get_apni(x="Acacia", rows = 1:10)

# When not found
get_apni("howdy")
get_apni(c("Acacia acicularis", "howdy"))

# Narrow down results 
# to accepted names
get_apni("Acacia", accepted = TRUE)
# to a kingom 
get_apni("Acacia", rank_filter = "genus")
# to accepted names and rank
get_apni("Acacia acicularis", accepted = TRUE, rank_filter = "species")
# to a family
get_apni("flower", family_filter = "Acanthaceae")

# Convert a apni without class information to a apni class
z <- get_apni("Acacia acicularis", accepted = TRUE, rank_filter = "species")
# already a apni, returns the same
as.apni(z)
as.apni("urn:lsid:ipni.org:names:119003-2")
# character vector, length > 1
ids <- c("urn:lsid:ipni.org:names:119003-2","urn:lsid:ipni.org:names:328247-2")
as.apni(ids)
# list, with character strings
as.apni(as.list(ids)) 
## dont check, much faster
as.apni("urn:lsid:ipni.org:names:119003-2", check=FALSE)
as.apni(ids, check=FALSE)
as.apni(as.list(ids), check=FALSE)

(out <- as.apni(ids))
data.frame(out)
as.apni( data.frame(out) )

# Get all data back
get_apni_("Quercus", rows=1:5)
get_apni_("Quercus", rows=1)
get_apni_(c("Pinus", "Abies"), rows = 1:3)
}
}
\seealso{
\code{\link[=classification]{classification()}}

Other apni: 
\code{\link{apni}}

Other taxonomic-ids: 
\code{\link{get_boldid}()},
\code{\link{get_eolid}()},
\code{\link{get_gbifid}()},
\code{\link{get_ids}()},
\code{\link{get_iucn}()},
\code{\link{get_natservid}()},
\code{\link{get_nbnid}()},
\code{\link{get_pow}()},
\code{\link{get_tolid}()},
\code{\link{get_tpsid}()},
\code{\link{get_tsn}()},
\code{\link{get_uid}()},
\code{\link{get_wiki}()},
\code{\link{get_wormsid}()}
}
\concept{apni}
\concept{taxonomic-ids}
