% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_boldid}
\alias{as.boldid}
\alias{as.boldid.boldid}
\alias{as.boldid.character}
\alias{as.boldid.data.frame}
\alias{as.boldid.list}
\alias{as.boldid.numeric}
\alias{as.data.frame.boldid}
\alias{get_boldid}
\alias{get_boldid_}
\title{Get the BOLD (Barcode of Life) code for a search term.}
\usage{
get_boldid(searchterm, fuzzy = FALSE, dataTypes = "basic",
  includeTree = FALSE, ask = TRUE, verbose = TRUE)

as.boldid(x)

\method{as.boldid}{boldid}(x)

\method{as.boldid}{character}(x)

\method{as.boldid}{list}(x)

\method{as.boldid}{numeric}(x)

\method{as.boldid}{data.frame}(x)

\method{as.data.frame}{boldid}(x, ...)

get_boldid_(searchterm, verbose = TRUE, fuzzy = FALSE,
  dataTypes = "basic", includeTree = FALSE, rows = NA, ...)
}
\arguments{
\item{searchterm}{character; A vector of common or scientific names.}

\item{fuzzy}{(logical) Whether to use fuzzy search or not (default: FALSE).}

\item{dataTypes}{(character) Specifies the datatypes that will be returned. See Details for
options.}

\item{includeTree}{(logical) If TRUE (default: FALSE), returns a list containing information
for parent taxa as well as the specified taxon.}

\item{ask}{logical; should get_tsn be run in interactive mode?
If TRUE and more than one TSN is found for teh species, the user is asked for
input. If FALSE NA is returned for multiple matches.}

\item{verbose}{logical; should progress be printed?}

\item{x}{Input to \code{\link{as.boldid}}}

\item{...}{Ignored}
}
\value{
A vector of BOLD ids. If a taxon is not found NA. If more than one BOLD ID is found
   the function asks for user input (if ask = TRUE), otherwise returns NA.
   Comes with an attribute \emph{match} to investigate the reason for NA (either 'not found',
   'found' or if ask = FALSE 'multi match')
}
\description{
Get the BOLD (Barcode of Life) code for a search term.
}
\examples{
\donttest{
get_boldid(searchterm = "Agapostemon")
get_boldid(searchterm = "Chironomus riparius")
get_boldid(c("Chironomus riparius","Quercus douglasii"))
splist <- names_list('species')
get_boldid(splist, verbose=FALSE)

# Fuzzy searching
get_boldid(searchterm="Osmi", fuzzy=TRUE)

# When not found
get_boldid("howdy")
get_boldid(c("Chironomus riparius", "howdy"))
get_boldid('Epicordulia princeps')
get_boldid('Arigomphus furcifer')
get_boldid("Cordulegaster erronea")
get_boldid("Nasiaeshna pentacantha")

# Convert a boldid without class information to a boldid class
as.boldid(get_boldid("Agapostemon")) # already a boldid, returns the same
as.boldid(get_boldid(c("Agapostemon","Quercus douglasii"))) # same
as.boldid(1973) # numeric
as.boldid(c(1973,101009,98597)) # numeric vector, length > 1
as.boldid("1973") # character
as.boldid(c("1973","101009","98597")) # character vector, length > 1
as.boldid(list("1973","101009","98597")) # list, either numeric or character

(out <- as.boldid(c(1973,101009,98597)))
data.frame(out)
as.boldid( data.frame(out) )

# Get all data back
get_boldid_("Osmia", fuzzy=TRUE, rows=1:5)
get_boldid_("Osmia", fuzzy=TRUE, rows=1)
}
}
\seealso{
\code{\link[taxize]{get_uid}}, \code{\link[taxize]{classification}}
}

