\name{getacceptednamesfromtsn}
\alias{getacceptednamesfromtsn}
\title{match count}
\usage{
  getacceptednamesfromtsn(tsn = NA, supmess = TRUE,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getAcceptedNamesFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

