% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getacceptednamesfromtsn}
\alias{getacceptednamesfromtsn}
\title{Get accepted names from tsn}
\usage{
getacceptednamesfromtsn(x, ...)
}
\arguments{
\item{x}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}

\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Get accepted names from tsn
}
\examples{
\donttest{
# TSN accepted - good name
library('httr')
getacceptednamesfromtsn('208527', config=timeout(1))

# TSN not accepted - input TSN is old name
getacceptednamesfromtsn('504239', config=timeout(1))

# TSN not accepted - input TSN is old name
getacceptednamesfromtsn('504239', config=timeout(3))
}
}
\keyword{internal}

