% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getanymatchcount}
\alias{getanymatchcount}
\title{Get any match count.}
\usage{
getanymatchcount(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\value{
An integer containing the number of matches the search will return.
}
\description{
Get any match count.
}
\examples{
\dontrun{
library('httr')
getanymatchcount(202385, config=timeout(3))
getanymatchcount("dolphin", config=timeout(3))
}
}
\keyword{internal}

