% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcommentdetailfromtsn}
\alias{getcommentdetailfromtsn}
\title{Get comment detail from TSN}
\usage{
getcommentdetailfromtsn(x, ...)
}
\arguments{
\item{x}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\value{
A data.frame with results.
}
\description{
Get comment detail from TSN
}
\examples{
\donttest{
getcommentdetailfromtsn(180543, config=timeout(1))
}
}
\keyword{internal}

