% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getcommentdetailfromtsn}
\alias{getcommentdetailfromtsn}
\title{Get comment detail from TSN}
\usage{
getcommentdetailfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle())
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\value{
A data.frame with results.
}
\description{
Get comment detail from TSN
}
\examples{
\donttest{
getcommentdetailfromtsn(180543)
}
}
\keyword{internal}

