\name{getcommentdetailfromtsn}
\alias{getcommentdetailfromtsn}
\title{match count}
\usage{
  getcommentdetailfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getCommentDetailFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

