% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcommonnamesfromtsn}
\alias{getcommonnamesfromtsn}
\title{Get common names from tsn}
\usage{
getcommonnamesfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get common names from tsn
}
\examples{
\donttest{
getcommonnamesfromtsn(183833, config=timeout(1))
}
}
\keyword{internal}

