% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcommonnamesfromtsn}
\alias{getcommonnamesfromtsn}
\title{Get common names from tsn}
\usage{
getcommonnamesfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get common names from tsn
}
\examples{
\dontrun{
getcommonnamesfromtsn(183833, config=timeout(1))
}
}
\keyword{internal}

