% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getcommonnamesfromtsn}
\alias{getcommonnamesfromtsn}
\title{Get common names from tsn}
\usage{
getcommonnamesfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get common names from tsn
}
\examples{
\dontrun{
getcommonnamesfromtsn(tsn=183833)
}
}
\keyword{internal}

