% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getcoremetadatafromtsn}
\alias{getcoremetadatafromtsn}
\title{Get core metadata from tsn}
\usage{
getcoremetadatafromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get core metadata from tsn
}
\examples{
\dontrun{
getcoremetadatafromtsn(tsn=28727)  # coverage and currrency data
getcoremetadatafromtsn(183671)  # no coverage or currrency data
}
}
\keyword{internal}

