% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcoremetadatafromtsn}
\alias{getcoremetadatafromtsn}
\title{Get core metadata from tsn}
\usage{
getcoremetadatafromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get core metadata from tsn
}
\examples{
\donttest{
getcoremetadatafromtsn(28727, config=timeout(3))  # coverage and currrency data
getcoremetadatafromtsn(183671, config=timeout(4))  # no coverage or currrency data
}
}
\keyword{internal}

