% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcoveragefromtsn}
\alias{getcoveragefromtsn}
\title{Get coverge from tsn}
\usage{
getcoveragefromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get coverge from tsn
}
\examples{
\dontrun{
getcoveragefromtsn(tsn=28727, config=timeout(4))  # coverage data
getcoveragefromtsn(526852, config=timeout(4))  # no coverage data
}
}
\keyword{internal}

