% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getcoveragefromtsn}
\alias{getcoveragefromtsn}
\title{Get coverge from tsn}
\usage{
getcoveragefromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get coverge from tsn
}
\examples{
\dontrun{
getcoveragefromtsn(tsn=28727)  # coverage data
getcoveragefromtsn(526852)  # no coverage data
}
}
\keyword{internal}

