\name{getcredibilityratingfromtsn}
\alias{getcredibilityratingfromtsn}
\title{match count}
\usage{
  getcredibilityratingfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getCredibilityRatingFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

