% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getcredibilityratingfromtsn}
\alias{getcredibilityratingfromtsn}
\title{Get credibility rating from tsn}
\usage{
getcredibilityratingfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get credibility rating from tsn
}
\examples{
\dontrun{
getcredibilityratingfromtsn(526852)
}
}
\keyword{internal}

