% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcredibilityratingfromtsn}
\alias{getcredibilityratingfromtsn}
\title{Get credibility rating from tsn}
\usage{
getcredibilityratingfromtsn(x, ...)
}
\arguments{
\item{x}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get credibility rating from tsn
}
\examples{
\donttest{
getcredibilityratingfromtsn(526852, config=timeout(4))
}
}
\keyword{internal}

