% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcredibilityratings}
\alias{getcredibilityratings}
\title{Get possible credibility ratings}
\usage{
getcredibilityratings()
}
\description{
Get possible credibility ratings
}
\examples{
\donttest{
getcredibilityratings()
}
}
\keyword{internal}

