% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcredibilityratings}
\alias{getcredibilityratings}
\title{Get possible credibility ratings}
\usage{
getcredibilityratings(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible credibility ratings
}
\examples{
\dontrun{
getcredibilityratings(config=timeout(3))
}
}
\keyword{internal}

