% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcurrencyfromtsn}
\alias{getcurrencyfromtsn}
\title{Get currency from tsn}
\usage{
getcurrencyfromtsn(x, ...)
}
\arguments{
\item{x}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get currency from tsn
}
\examples{
\donttest{
getcurrencyfromtsn(28727, config=timeout(3)) # currency data
getcurrencyfromtsn(x=526852, config=timeout(3)) # no currency dat
}
}
\keyword{internal}

