% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getcurrencyfromtsn}
\alias{getcurrencyfromtsn}
\title{Get currency from tsn}
\usage{
getcurrencyfromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle())
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Get currency from tsn
}
\examples{
\donttest{
getcurrencyfromtsn(tsn = 28727) # currency data
getcurrencyfromtsn(tsn = 526852) # no currency dat
}
}
\keyword{internal}

