% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getcurrencyfromtsn}
\alias{getcurrencyfromtsn}
\title{Get currency from tsn}
\usage{
getcurrencyfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get currency from tsn
}
\examples{
\dontrun{
getcurrencyfromtsn(28727, config=timeout(3)) # currency data
getcurrencyfromtsn(526852, config=timeout(3)) # no currency dat
}
}
\keyword{internal}

