% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getdatedatafromtsn}
\alias{getdatedatafromtsn}
\title{Get date data from tsn}
\usage{
getdatedatafromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get date data from tsn
}
\examples{
\dontrun{
getdatedatafromtsn(180543, config=timeout(3))
}
}
\keyword{internal}

