% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getdescription}
\alias{getdescription}
\title{Get description of the ITIS service}
\usage{
getdescription(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get description of the ITIS service
}
\examples{
\dontrun{
getdescription(config=timeout(1))
}
}
\keyword{internal}

