% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getexpertsfromtsn}
\alias{getexpertsfromtsn}
\title{Get expert information for the TSN.}
\usage{
getexpertsfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get expert information for the TSN.
}
\examples{
\dontrun{
getexpertsfromtsn(180544, config=timeout(3))
}
}
\keyword{internal}

