% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getexpertsfromtsn}
\alias{getexpertsfromtsn}
\title{Get expert information for the TSN.}
\usage{
getexpertsfromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle())
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Get expert information for the TSN.
}
\examples{
\donttest{
getexpertsfromtsn(tsn = 180544)
}
}
\keyword{internal}

