\name{getfullhierarchyfromtsn}
\alias{getfullhierarchyfromtsn}
\title{match count}
\usage{
  getfullhierarchyfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getFullHierarchyFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

