% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getfullhierarchyfromtsn}
\alias{getfullhierarchyfromtsn}
\title{Get full hierarchy from tsn}
\usage{
getfullhierarchyfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get full hierarchy from tsn
}
\examples{
\dontrun{
getfullhierarchyfromtsn(37906, config=timeout(3))
getfullhierarchyfromtsn(100800, config=timeout(3))
}
}
\keyword{internal}

