% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getfullrecordfromlsid}
\alias{getfullrecordfromlsid}
\title{Returns the full ITIS record for the TSN in the LSID, found by comparing the
		TSN in the search key to the TSN field. Returns an empty result set if
  	there is no match or the TSN is invalid.}
\usage{
getfullrecordfromlsid(lsid, ...)
}
\arguments{
\item{lsid}{lsid for a taxonomic group (character)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the full ITIS record for the TSN in the LSID, found by comparing the
		TSN in the search key to the TSN field. Returns an empty result set if
  	there is no match or the TSN is invalid.
}
\examples{
\dontrun{
getfullrecordfromlsid("urn:lsid:itis.gov:itis_tsn:180543", config=timeout(3))
getfullrecordfromlsid("urn:lsid:itis.gov:itis_tsn:180543", config=timeout(3))
}
}
\keyword{internal}

