% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getfullrecordfromtsn}
\alias{getfullrecordfromtsn}
\title{Get full record from TSN.}
\usage{
getfullrecordfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get full record from TSN.
}
\examples{
\dontrun{
getfullrecordfromtsn(504239, config=timeout(3))
getfullrecordfromtsn(202385, config=timeout(3))
getfullrecordfromtsn(183833, config=timeout(3))
}
}
\keyword{internal}

