% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getfullrecordfromtsn}
\alias{getfullrecordfromtsn}
\title{Get full record from TSN.}
\usage{
getfullrecordfromtsn(x, ...)
}
\arguments{
\item{x}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}

\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Get full record from TSN.
}
\examples{
\donttest{
getfullrecordfromtsn(504239, config=timeout(3))
getfullrecordfromtsn(202385, config=timeout(3))
getfullrecordfromtsn(183833, config=timeout(3))
getfullrecordfromtsn(183833, config=timeout(3))
}
}
\keyword{internal}

