% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getgeographicdivisionsfromtsn}
\alias{getgeographicdivisionsfromtsn}
\title{Get geographic divisions from tsn}
\usage{
getgeographicdivisionsfromtsn(x, ...)
}
\arguments{
\item{x}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get geographic divisions from tsn
}
\examples{
\donttest{
getgeographicdivisionsfromtsn(180543, config=timeout(3))
}
}
\keyword{internal}

