% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getgeographicdivisionsfromtsn}
\alias{getgeographicdivisionsfromtsn}
\title{Get geographic divisions from tsn}
\usage{
getgeographicdivisionsfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get geographic divisions from tsn
}
\examples{
\dontrun{
getgeographicdivisionsfromtsn(tsn=180543, config=timeout(3))
}
}
\keyword{internal}

