% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getgeographicvalues}
\alias{getgeographicvalues}
\title{Get all possible geographic values}
\usage{
getgeographicvalues()
}
\description{
Get all possible geographic values
}
\examples{
\dontrun{
getgeographicvalues()
}
}
\keyword{internal}

