% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getgeographicvalues}
\alias{getgeographicvalues}
\title{Get all possible geographic values}
\usage{
getgeographicvalues(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get all possible geographic values
}
\examples{
\dontrun{
getgeographicvalues()
}
}
\keyword{internal}

