\name{getglobalspeciescompletenessfromtsn}
\alias{getglobalspeciescompletenessfromtsn}
\title{match count}
\usage{
  getglobalspeciescompletenessfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getGlobalSpeciesCompletenessFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

