% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getglobalspeciescompletenessfromtsn}
\alias{getglobalspeciescompletenessfromtsn}
\title{Get global species completeness from tsn}
\usage{
getglobalspeciescompletenessfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get global species completeness from tsn
}
\examples{
\dontrun{
getglobalspeciescompletenessfromtsn(tsn=180541, config=timeout(3))
}
}
\keyword{internal}

