\name{gethierarchydownfromtsn}
\alias{gethierarchydownfromtsn}
\title{match count}
\usage{
  gethierarchydownfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getHierarchyDownFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

