% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{gethierarchydownfromtsn}
\alias{gethierarchydownfromtsn}
\title{Get hierarchy down from tsn}
\usage{
gethierarchydownfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get hierarchy down from tsn
}
\examples{
\dontrun{
gethierarchydownfromtsn(161030, config=timeout(3))
}
}
\keyword{internal}

