% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gethierarchydownfromtsn}
\alias{gethierarchydownfromtsn}
\title{Get hierarchy down from tsn}
\usage{
gethierarchydownfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}

\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Get hierarchy down from tsn
}
\examples{
\donttest{
gethierarchydownfromtsn(tsn = 161030)
}
}
\keyword{internal}

