% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gethierarchyupfromtsn}
\alias{gethierarchyupfromtsn}
\title{Get hierarchy up from tsn}
\usage{
gethierarchyupfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get hierarchy up from tsn
}
\examples{
\dontrun{
gethierarchyupfromtsn(36485, config=timeout(3))
}
}
\keyword{internal}

