% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getitisterms}
\alias{getitisterms}
\title{Get itis terms from common names}
\usage{
getitisterms(srchkey = NA, curlopts = list(), curl = getCurlHandle(),
  verbose = TRUE)
}
\arguments{
\item{srchkey}{text or taxonomic serial number (TSN) (character or numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Get itis terms from common names
}
\examples{
\donttest{
getitisterms(srchkey="bear")
}
}
\keyword{internal}

