\name{getitistermsfromcommonname}
\alias{getitistermsfromcommonname}
\title{match count}
\usage{
  getitistermsfromcommonname(srchkey = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getITISTermsFromCommonName",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

