% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{getitistermsfromcommonname}
\alias{getitistermsfromcommonname}
\title{Get itis terms from common names}
\usage{
getitistermsfromcommonname(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get itis terms from common names
}
\examples{
\dontrun{
getitistermsfromcommonname("buya")

library('httr')
getitistermsfromcommonname("buya", config=timeout(1))
}
}
\keyword{internal}

