% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getitistermsfromcommonname}
\alias{getitistermsfromcommonname}
\title{Get itis terms from common names}
\usage{
getitistermsfromcommonname(srchkey = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{srchkey}{text or taxonomic serial number (TSN) (character or numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Get itis terms from common names
}
\examples{
\donttest{
getitistermsfromcommonname("buya")
}
}
\keyword{internal}

