\name{getitistermsfromscientificname}
\alias{getitistermsfromscientificname}
\title{match count}
\usage{
  getitistermsfromscientificname(srchkey = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getITISTermsFromScientificName",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

