% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getitistermsfromscientificname}
\alias{getitistermsfromscientificname}
\title{Get itis terms from scientific names}
\usage{
getitistermsfromscientificname(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get itis terms from scientific names
}
\examples{
\dontrun{
getitistermsfromscientificname(x="ursidae")
getitistermsfromscientificname("Ursus")
}
}
\keyword{internal}

