% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getitistermsfromscientificname}
\alias{getitistermsfromscientificname}
\title{Get itis terms from scientific names}
\usage{
getitistermsfromscientificname(srchkey = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{srchkey}{text or taxonomic serial number (TSN) (character or numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Get itis terms from scientific names
}
\examples{
\donttest{
getitistermsfromscientificname(srchkey = "ursidae")
getitistermsfromscientificname(srchkey = "Ursus")
}
}
\keyword{internal}

