\name{getjurisdictionaloriginfromtsn}
\alias{getjurisdictionaloriginfromtsn}
\title{match count}
\usage{
  getjurisdictionaloriginfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getJurisdictionalOriginFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

