% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getjurisdictionaloriginfromtsn}
\alias{getjurisdictionaloriginfromtsn}
\title{Get jurisdictional origin from tsn}
\usage{
getjurisdictionaloriginfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{verbose}{Verbosity or not (default TRUE)}

\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Get jurisdictional origin from tsn
}
\examples{
\donttest{
getjurisdictionaloriginfromtsn(tsn = 2180543)
}
}
\keyword{internal}

