% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getjurisdictionaloriginfromtsn}
\alias{getjurisdictionaloriginfromtsn}
\title{Get jurisdictional origin from tsn}
\usage{
getjurisdictionaloriginfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get jurisdictional origin from tsn
}
\examples{
\donttest{
getjurisdictionaloriginfromtsn(2180543, config=timeout(3))
}
}
\keyword{internal}

