% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getjurisdictionaloriginfromtsn}
\alias{getjurisdictionaloriginfromtsn}
\title{Get jurisdictional origin from tsn}
\usage{
getjurisdictionaloriginfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Get jurisdictional origin from tsn
}
\examples{
\donttest{
getjurisdictionaloriginfromtsn(tsn = 2180543)
}
}
\keyword{internal}

