% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getjurisdictionoriginvalues}
\alias{getjurisdictionoriginvalues}
\title{Get jurisdiction origin values}
\usage{
getjurisdictionoriginvalues()
}
\description{
Get jurisdiction origin values
}
\examples{
\dontrun{
getjurisdictionoriginvalues()
}
}
\keyword{internal}

