% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getjurisdictionoriginvalues}
\alias{getjurisdictionoriginvalues}
\title{Get jurisdiction origin values}
\usage{
getjurisdictionoriginvalues(verbose = TRUE)
}
\description{
Get jurisdiction origin values
}
\examples{
\donttest{
getjurisdictionoriginvalues()
}
}
\keyword{internal}

