% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getjurisdictionoriginvalues}
\alias{getjurisdictionoriginvalues}
\title{Get jurisdiction origin values}
\usage{
getjurisdictionoriginvalues(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get jurisdiction origin values
}
\examples{
\dontrun{
getjurisdictionoriginvalues(config=timeout(3))
}
}
\keyword{internal}

