% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getjurisdictionvalues}
\alias{getjurisdictionvalues}
\title{Get possible jurisdiction values}
\usage{
getjurisdictionvalues(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible jurisdiction values
}
\examples{
\donttest{
getjurisdictionvalues(config=timeout(3))
}
}
\keyword{internal}

