% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getjurisdictionvalues}
\alias{getjurisdictionvalues}
\title{Get possible jurisdiction values}
\usage{
getjurisdictionvalues(verbose = TRUE)
}
\description{
Get possible jurisdiction values
}
\examples{
\donttest{
getjurisdictionvalues()
}
}
\keyword{internal}

