% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getjurisdictionvalues}
\alias{getjurisdictionvalues}
\title{Get possible jurisdiction values}
\usage{
getjurisdictionvalues(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible jurisdiction values
}
\examples{
\dontrun{
getjurisdictionvalues(config=timeout(3))
}
}
\keyword{internal}

