% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getkey.R
\name{getkey}
\alias{getkey}
\title{Function to get API key.}
\usage{
getkey(x = NULL, service)
}
\arguments{
\item{x}{(character) An API key, defaults to \code{NULL}}

\item{service}{(character) The API data provider, used to match to
default guest key (for Tropicos; there's no guest
key for NCBI or IUCN, for which you have to get your own)}
}
\description{
Checks first to get key from your .Rprofile or .Renviron (or similar) file
}
\examples{
\dontrun{
getkey(service="tropicos")
getkey(service="iucn")
getkey(service="entrez")
}
}
