% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getkingdomnamefromtsn}
\alias{getkingdomnamefromtsn}
\title{Get kingdom names from tsn}
\usage{
getkingdomnamefromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle(),
  verbose = TRUE)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}

\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Get kingdom names from tsn
}
\examples{
\donttest{
getkingdomnamefromtsn(tsn = 202385)
}
}
\keyword{internal}

