% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getkingdomnamefromtsn}
\alias{getkingdomnamefromtsn}
\title{Get kingdom names from tsn}
\usage{
getkingdomnamefromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle(),
  verbose = TRUE)
}
\arguments{
\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Get kingdom names from tsn
}
\examples{
\donttest{
getkingdomnamefromtsn(tsn = 202385)
}
}
\keyword{internal}

