% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getkingdomnamefromtsn}
\alias{getkingdomnamefromtsn}
\title{Get kingdom names from tsn}
\usage{
getkingdomnamefromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Get kingdom names from tsn
}
\examples{
\dontrun{
getkingdomnamefromtsn(202385, config=timeout(3))
}
}
\keyword{internal}

