% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getkingdomnames}
\alias{getkingdomnames}
\title{Get all possible kingdom names}
\usage{
getkingdomnames(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get all possible kingdom names
}
\examples{
\dontrun{
getkingdomnames(config=timeout(3))
}
}
\keyword{internal}

