% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getkingdomnames}
\alias{getkingdomnames}
\title{Get all possible kingdom names}
\usage{
getkingdomnames(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get all possible kingdom names
}
\examples{
\donttest{
getkingdomnames(config=timeout(3))
}
}
\keyword{internal}

