% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getlastchangedate}
\alias{getlastchangedate}
\title{Provides the date the ITIS database was last updated.}
\usage{
getlastchangedate(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Provides the date the ITIS database was last updated.
}
\examples{
\dontrun{
getlastchangedate(config=timeout(3))
}
}
\keyword{internal}

