% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getlastchangedate}
\alias{getlastchangedate}
\title{Provides the date the ITIS database was last updated.}
\usage{
getlastchangedate()
}
\description{
Provides the date the ITIS database was last updated.
}
\examples{
\dontrun{
getlastchangedate()
}
}
\keyword{internal}

