% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getlsidfromtsn}
\alias{getlsidfromtsn}
\title{Gets the unique LSID for the TSN, or an empty result if there is no match.}
\usage{
getlsidfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Gets the unique LSID for the TSN, or an empty result if there is no match.
}
\examples{
\dontrun{
getlsidfromtsn(155166, config=timeout(3))
}
}
\keyword{internal}

