% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getothersourcesfromtsn}
\alias{getothersourcesfromtsn}
\title{Returns a list of the other sources used for the TSN.}
\usage{
getothersourcesfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle())
}
\description{
Returns a list of the other sources used for the TSN.
}
\examples{
\donttest{
getothersourcesfromtsn(tsn = 182662)
}
}
\keyword{internal}

