\name{getothersourcesfromtsn}
\alias{getothersourcesfromtsn}
\title{match count}
\usage{
  getothersourcesfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getOtherSourcesFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

