% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getothersourcesfromtsn}
\alias{getothersourcesfromtsn}
\title{Returns a list of the other sources used for the TSN.}
\usage{
getothersourcesfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns a list of the other sources used for the TSN.
}
\examples{
\donttest{
getothersourcesfromtsn(182662, config=timeout(3))
}
}
\keyword{internal}

