% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getparenttsnfromtsn}
\alias{getparenttsnfromtsn}
\title{Returns the parent TSN for the entered TSN.}
\usage{
getparenttsnfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the parent TSN for the entered TSN.
}
\examples{
\dontrun{
getparenttsnfromtsn(202385, config=timeout(3))
}
}
\keyword{internal}

