% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getparenttsnfromtsn}
\alias{getparenttsnfromtsn}
\title{Returns the parent TSN for the entered TSN.}
\usage{
getparenttsnfromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle())
}
\description{
Returns the parent TSN for the entered TSN.
}
\examples{
\donttest{
getparenttsnfromtsn(tsn = 202385)
}
}
\keyword{internal}

