\name{getpublicationsfromtsn}
\alias{getpublicationsfromtsn}
\title{match count}
\usage{
  getpublicationsfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getPublicationsFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

