% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getpublicationsfromtsn}
\alias{getpublicationsfromtsn}
\title{Returns a list of the pulications used for the TSN.}
\usage{
getpublicationsfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns a list of the pulications used for the TSN.
}
\examples{
\donttest{
getpublicationsfromtsn(70340, config=timeout(3))
}
}
\keyword{internal}

