% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getranknames}
\alias{getranknames}
\title{Provides a list of all the unique rank names contained in the database and
 their kingdom and rank ID values.}
\usage{
getranknames(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Provides a list of all the unique rank names contained in the database and
 their kingdom and rank ID values.
}
\examples{
\dontrun{
getranknames(config=timeout(3))
}
}
\keyword{internal}

