\name{getranknames}
\alias{getranknames}
\title{Provides a list of all the unique rank names contained in the database and
 their kingdom and rank ID values.}
\usage{
  getranknames()
}
\description{
  Provides a list of all the unique rank names contained in
  the database and their kingdom and rank ID values.
}
\examples{
\dontrun{
getranknames()
}
}
\keyword{internal}

