% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getrecordfromlsid}
\alias{getrecordfromlsid}
\title{Gets the partial ITIS record for the TSN in the LSID, found by comparing the
 TSN in the search key to the TSN field. Returns an empty result set if
 there is no match or the TSN is invalid.}
\usage{
getrecordfromlsid(lsid, ...)
}
\arguments{
\item{lsid}{lsid for a taxonomic group (character)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Gets the partial ITIS record for the TSN in the LSID, found by comparing the
 TSN in the search key to the TSN field. Returns an empty result set if
 there is no match or the TSN is invalid.
}
\examples{
\dontrun{
getrecordfromlsid("urn:lsid:itis.gov:itis_tsn:180543", config=timeout(5))
}
}
\keyword{internal}

