% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getreviewyearfromtsn}
\alias{getreviewyearfromtsn}
\title{Returns the review year for the TSN.}
\usage{
getreviewyearfromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle())
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Returns the review year for the TSN.
}
\examples{
\donttest{
getreviewyearfromtsn(tsn = 180541)
}
}
\keyword{internal}

