% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getreviewyearfromtsn}
\alias{getreviewyearfromtsn}
\title{Returns the review year for the TSN.}
\usage{
getreviewyearfromtsn(tsn = NA, curlopts = list(), curl = getCurlHandle())
}
\description{
Returns the review year for the TSN.
}
\examples{
\donttest{
getreviewyearfromtsn(tsn = 180541)
}
}
\keyword{internal}

