% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis.R
\name{getreviewyearfromtsn}
\alias{getreviewyearfromtsn}
\title{Returns the review year for the TSN.}
\usage{
getreviewyearfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the review year for the TSN.
}
\examples{
\dontrun{
getreviewyearfromtsn(180541, config=timeout(3))
}
}
\keyword{internal}

