% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getscientificnamefromtsn}
\alias{getscientificnamefromtsn}
\title{Returns the scientific name for the TSN. Also returns the component parts
   (names and indicators) of the scientific name.}
\usage{
getscientificnamefromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle())
}
\description{
Returns the scientific name for the TSN. Also returns the component parts
   (names and indicators) of the scientific name.
}
\examples{
\donttest{
getscientificnamefromtsn(tsn = 531894)
}
}
\keyword{internal}

