\name{getscientificnamefromtsn}
\alias{getscientificnamefromtsn}
\title{match count}
\usage{
  getscientificnamefromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getScientificNameFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

