% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getscientificnamefromtsn}
\alias{getscientificnamefromtsn}
\title{Returns the scientific name for the TSN. Also returns the component parts
   (names and indicators) of the scientific name.}
\usage{
getscientificnamefromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the scientific name for the TSN. Also returns the component parts
   (names and indicators) of the scientific name.
}
\examples{
\dontrun{
getscientificnamefromtsn(531894, config=timeout(3))
}
}
\keyword{internal}

