\name{getsynonymnamesfromtsn}
\alias{getsynonymnamesfromtsn}
\title{match count}
\usage{
  getsynonymnamesfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getSynonymNamesFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

