% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{getsynonymnamesfromtsn}
\alias{getsynonymnamesfromtsn}
\title{Returns a list of the synonyms (if any) for the TSN.}
\usage{
getsynonymnamesfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns a list of the synonyms (if any) for the TSN.
}
\examples{
\dontrun{
getsynonymnamesfromtsn(183671, config=timeout(3)) # tsn not accepted
getsynonymnamesfromtsn(526852, config=timeout(5)) # tsn accepted
}
}
\keyword{internal}

