% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getsynonymnamesfromtsn}
\alias{getsynonymnamesfromtsn}
\title{Returns a list of the synonyms (if any) for the TSN.}
\usage{
getsynonymnamesfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}

\item{verbose}{(logical) Print messages or not. Default to TRUE.}
}
\description{
Returns a list of the synonyms (if any) for the TSN.
}
\examples{
\donttest{
getsynonymnamesfromtsn(tsn = 183671) # tsn not accepted
getsynonymnamesfromtsn(tsn = 526852) # tsn accepted
}
}
\keyword{internal}

