\name{gettaxonauthorshipfromtsn}
\alias{gettaxonauthorshipfromtsn}
\title{match count}
\usage{
  gettaxonauthorshipfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getTaxonAuthorshipFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

