% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettaxonauthorshipfromtsn}
\alias{gettaxonauthorshipfromtsn}
\title{Returns the author information for the TSN.}
\usage{
gettaxonauthorshipfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle())
}
\description{
Returns the author information for the TSN.
}
\examples{
\donttest{
gettaxonauthorshipfromtsn(tsn = 183671)
}
}
\keyword{internal}

