% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gettaxonauthorshipfromtsn}
\alias{gettaxonauthorshipfromtsn}
\title{Returns the author information for the TSN.}
\usage{
gettaxonauthorshipfromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle())
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Returns the author information for the TSN.
}
\examples{
\dontrun{
gettaxonauthorshipfromtsn(tsn = 183671)
}
}
\keyword{internal}

