% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettaxonauthorshipfromtsn}
\alias{gettaxonauthorshipfromtsn}
\title{Returns the author information for the TSN.}
\usage{
gettaxonauthorshipfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the author information for the TSN.
}
\examples{
\donttest{
gettaxonauthorshipfromtsn(183671, config=timeout(3))
}
}
\keyword{internal}

