% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{gettaxonauthorshipfromtsn}
\alias{gettaxonauthorshipfromtsn}
\title{Returns the author information for the TSN.}
\usage{
gettaxonauthorshipfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the author information for the TSN.
}
\examples{
\dontrun{
gettaxonauthorshipfromtsn(183671, config=timeout(3))
}
}
\keyword{internal}

