% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettaxonomicranknamefromtsn}
\alias{gettaxonomicranknamefromtsn}
\title{Returns the kingdom and rank information for the TSN.}
\usage{
gettaxonomicranknamefromtsn(tsn = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{verbose}{Verbosity or not (default TRUE)}
}
\description{
Returns the kingdom and rank information for the TSN.
}
\examples{
\donttest{
gettaxonomicranknamefromtsn(tsn = 202385)
}
}
\keyword{internal}

