% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettaxonomicranknamefromtsn}
\alias{gettaxonomicranknamefromtsn}
\title{Returns the kingdom and rank information for the TSN.}
\usage{
gettaxonomicranknamefromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the kingdom and rank information for the TSN.
}
\examples{
\donttest{
gettaxonomicranknamefromtsn(202385, config=timeout(3))
}
}
\keyword{internal}

