\name{gettaxonomicranknamefromtsn}
\alias{gettaxonomicranknamefromtsn}
\title{match count}
\usage{
  gettaxonomicranknamefromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getTaxonomicRankNameFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

