\name{gettaxonomicusagefromtsn}
\alias{gettaxonomicusagefromtsn}
\title{match count}
\usage{
  gettaxonomicusagefromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getTaxonomicUsageFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

