\name{gettsnbyvernacularlanguage}
\alias{gettsnbyvernacularlanguage}
\title{match count}
\usage{
  gettsnbyvernacularlanguage(language = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getTsnByVernacularLanguage",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

