% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gettsnbyvernacularlanguage}
\alias{gettsnbyvernacularlanguage}
\title{Get tsn by vernacular language}
\usage{
gettsnbyvernacularlanguage(language, ...)
}
\arguments{
\item{language}{A string containing the language. This is a language string,
not the international language code (character)}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get tsn by vernacular language
}
\examples{
\dontrun{
gettsnbyvernacularlanguage("french", config=timeout(3))
}
}
\keyword{internal}

