\name{getunacceptabilityreasonfromtsn}
\alias{getunacceptabilityreasonfromtsn}
\title{match count}
\usage{
  getunacceptabilityreasonfromtsn(tsn = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/getUnacceptabilityReasonFromTSN",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

