% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getunacceptabilityreasonfromtsn}
\alias{getunacceptabilityreasonfromtsn}
\title{Returns the unacceptability reason, if any, for the TSN.}
\usage{
getunacceptabilityreasonfromtsn(tsn, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Returns the unacceptability reason, if any, for the TSN.
}
\examples{
\dontrun{
getunacceptabilityreasonfromtsn(183671, config=timeout(3))
}
}
\keyword{internal}

