% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getvernacularlanguages}
\alias{getvernacularlanguages}
\title{Provides a list of the unique languages used in the vernacular table.}
\usage{
getvernacularlanguages()
}
\description{
Provides a list of the unique languages used in the vernacular table.
}
\examples{
\donttest{
getvernacularlanguages()
}
}
\keyword{internal}

