% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getvernacularlanguages}
\alias{getvernacularlanguages}
\title{Provides a list of the unique languages used in the vernacular table.}
\usage{
getvernacularlanguages(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Provides a list of the unique languages used in the vernacular table.
}
\examples{
\dontrun{
getvernacularlanguages(config=timeout(3))
}
}
\keyword{internal}

