\name{getvernacularlanguages}
\alias{getvernacularlanguages}
\title{Provides a list of the unique languages used in the vernacular table.}
\usage{
  getvernacularlanguages()
}
\description{
  Provides a list of the unique languages used in the
  vernacular table.
}
\examples{
\dontrun{
getvernacularlanguages()
}
}

