% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ubio_classification.R
\name{ifelsedata}
\alias{ifelsedata}
\title{Function to parse xml data, decode strings, and return NULL if logical is zero.}
\usage{
ifelsedata(a, x, y, z)
}
\description{
Function to parse xml data, decode strings, and return NULL if logical is zero.
}
\keyword{internal}

