% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion.R
\name{ion}
\alias{ion}
\title{ION - Index to Organism Names}
\usage{
ion(x, ...)
}
\arguments{
\item{x}{An LSID number. Required.}

\item{...}{Curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
A \link[tibble:tibble]{tibble::tibble}
}
\description{
ION - Index to Organism Names
}
\examples{
\dontrun{
ion(155166)
ion(298678)
ion(4796748) # ursus americanus
ion(1280626) # puma concolor
}
}
\references{
http://www.organismnames.com
}
