% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_name.R
\name{itis_name}
\alias{itis_name}
\title{Get taxonomic names for a given taxonomic name query.}
\usage{
itis_name(query = NULL, get = NULL)
}
\arguments{
\item{query}{TSN number (taxonomic serial number).}

\item{get}{The rank of the taxonomic name to get.}
}
\value{
Taxonomic name for the searched taxon.
}
\description{
Get taxonomic names for a given taxonomic name query.
}
\examples{
\dontrun{
itis_name(query="Helianthus annuus", get="family")
}
}
