% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis_ping.R
\name{itis_ping}
\alias{itis_ping}
\title{Ping the ITIS API to see if it's working.}
\usage{
itis_ping(what = "status", ...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Ping the ITIS API to see if it's working.
}
\details{
See \code{\link{getdescription()}}, which provides number of scientific and
common names in a character string.
}
\examples{
\dontrun{
itis_ping()
}
}

