% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itis_ping.R
\name{itis_ping}
\alias{itis_ping}
\title{Ping the ITIS API to see if it's working.}
\usage{
itis_ping(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
And provides number of scientific and common names in a string.
}
\examples{
\dontrun{
itis_ping()
}
}
\seealso{
\code{\link{eol_ping}}
}

