% Generated by roxygen2 (4.0.2): do not edit by hand
\name{itis_searchcommon}
\alias{itis_searchcommon}
\title{Searches common name and acts as thin wrapper around \code{searchbycommonnamebeginswith} and \code{searchbycommonnameendswith}}
\usage{
itis_searchcommon(srchkey = NA, curlopts = list(), curl = getCurlHandle(),
  from = "begin")
}
\arguments{
\item{srchkey}{search terms}

\item{curlopts}{Additional curl options.}

\item{curl}{curl handle}

\item{from}{default is to search from beginning. Use \code{end} to serch from end.}
}
\value{
\code{data.frame}
}
\description{
Searches common name and acts as thin wrapper around \code{searchbycommonnamebeginswith} and \code{searchbycommonnameendswith}
}
\examples{
\donttest{
searchCommon(srchkey="inch")
searchCommon(srchkey="inch", from = "end")
}
}
\seealso{
searchbycommonnamebeginswith searchbycommonnameendswith
}

