% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{itis_searchcommon}
\alias{itis_searchcommon}
\title{Searches common name and acts as thin wrapper around \code{searchbycommonnamebeginswith} and \code{searchbycommonnameendswith}}
\usage{
itis_searchcommon(x, from = "begin", ...)
}
\arguments{
\item{x}{Search terms}

\item{from}{Default is to search from beginning. Use \code{end} to serch from end.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
\code{data.frame}
}
\description{
Searches common name and acts as thin wrapper around \code{searchbycommonnamebeginswith} and \code{searchbycommonnameendswith}
}
\examples{
\dontrun{
itis_searchcommon(x="inch", config=timeout(3))
itis_searchcommon("inch", from = "end", config=timeout(3))
}
}
\seealso{
searchbycommonnamebeginswith searchbycommonnameendswith
}

