% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nbn_classifcation}
\alias{nbn_classifcation}
\title{Search UK National Biodiversity Network database for taxonomic classification}
\usage{
nbn_classifcation(id, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}}.}

\item{q}{(character) The query terms(s)}

\item{prefered}{(logical) Restrict search to preferred or any}

\item{order}{(character) The order in which we should sort the results. Default: asc}

\item{sort}{(character) Sort the results or not.}

\item{start}{(integer/numeric) The page that the user wants to start displaying the results at.
Default: 0}

\item{rows}{(integer/numeric) The number of rows to show in each page of search results.
Default: 25}

\item{taxonOutputGroupKey}{(character) Vector of taxon output groups.}

\item{all}{(logical) Get all results, overrides rows parameter if TRUE. Default: FALSE}
}
\description{
Search UK National Biodiversity Network database for taxonomic classification
}
\examples{
\donttest{
nbn_classifcation(id="NHMSYS0000502940")

# get id first, then pass to this fxn
id <- get_nbnid("blue tit", rec_only = TRUE, rank = "Species")
nbn_classifcation(id)

library('httr')
nbn_classifcation(id="NHMSYS0000502940", config=verbose())
}
}
\author{
Scott Chamberlain, \email{myrmecocystus@gmail.com}
}

