% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressor.R
\docType{data}
\name{progressor}
\alias{progressor}
\title{methods for preparing/printing info for prompts for \code{get_*} functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
progressor
}
\description{
methods for preparing/printing info for prompts for \code{get_*} functions
}
\examples{
\dontrun{
nms <- c("Quercus", "Sasdsfasdf")
x <- progressor$new(items = nms)
x
x$prog_start()

x$completed(nms[1], "found")
x$prog_found()

x$completed(nms[2], "not found")
x$prog_not_found()

x$prog_summary()

# suppress cli::cat_line
x <- progressor$new(items = nms, suppress = TRUE)
x$prog_summary()
}
}
\keyword{internal}
