% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{rank_ref}
\alias{rank_ref}
\title{Lookup-table for IDs of taxonomic ranks}
\description{
data.frame of 35 rows, with 2 columns:
\itemize{
 \item rankid - a numeric rank id, consecutive
 \item ranks - a comma separated vector of names that are equal to
 one another within the row
}
}
\keyword{data}
