% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\title{Resolve names from different data sources

DEFUNCT}
\usage{
resolve(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Resolve names from different data sources

DEFUNCT
}
