% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchbycommonname}
\alias{searchbycommonname}
\title{Search for tsn by common name}
\usage{
searchbycommonname(srchkey = NA, curlopts = list(),
  curl = getCurlHandle())
}
\description{
Search for tsn by common name
}
\examples{
\donttest{
searchbycommonname("american bullfrog")
searchbycommonname("ferret-badger")
searchbycommonname(srchkey="polar bear")
}
}
\keyword{internal}

