% Generated by roxygen2 (4.0.1): do not edit by hand
\name{searchbycommonname}
\alias{searchbycommonname}
\title{Search for tsn by common name}
\usage{
searchbycommonname(srchkey = NA, curlopts = list(),
  curl = getCurlHandle())
}
\arguments{
\item{srchkey}{text or taxonomic serial number (TSN) (character or numeric)}

\item{curlopts}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\description{
Search for tsn by common name
}
\examples{
\donttest{
searchbycommonname("american bullfrog")
searchbycommonname("ferret-badger")
searchbycommonname(srchkey="polar bear")
}
}
\keyword{internal}

