% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{searchbycommonname}
\alias{searchbycommonname}
\title{Search for tsn by common name}
\usage{
searchbycommonname(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for tsn by common name
}
\examples{
\dontrun{
searchbycommonname(x="american bullfrog", config=timeout(6))
searchbycommonname("ferret-badger", config=timeout(6))
searchbycommonname("polar bear", config=timeout(6))
}
}
\keyword{internal}

