\name{searchbycommonnamebeginswith}
\alias{searchbycommonnamebeginswith}
\title{match count}
\usage{
  searchbycommonnamebeginswith(srchkey = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/searchByCommonNameBeginsWith",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

