% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchbycommonnamebeginswith}
\alias{searchbycommonnamebeginswith}
\title{Search for tsn by common name beginning with}
\usage{
searchbycommonnamebeginswith(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for tsn by common name beginning with
}
\examples{
\donttest{
searchbycommonnamebeginswith("inch", config=timeout(3))
}
}
\keyword{internal}

