\name{searchbycommonnameendswith}
\alias{searchbycommonnameendswith}
\title{match count}
\usage{
  searchbycommonnameendswith(srchkey = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/searchByCommonNameEndsWith",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

