\name{searchbyscientificname}
\alias{searchbyscientificname}
\title{match count}
\usage{
  searchbyscientificname(srchkey = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/searchByScientificName",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

