% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchbyscientificname}
\alias{searchbyscientificname}
\title{Search by scientific name}
\usage{
searchbyscientificname(srchkey = NA, curlopts = list(),
  curl = getCurlHandle(), verbose = TRUE)
}
\description{
Search by scientific name
}
\examples{
\donttest{
searchbyscientificname("Tardigrada")
searchbyscientificname("Quercus douglasii")
}
}
\keyword{internal}

