% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itis.R
\name{searchbyscientificname}
\alias{searchbyscientificname}
\title{Search by scientific name}
\usage{
searchbyscientificname(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search by scientific name
}
\examples{
\dontrun{
searchbyscientificname("Tardigrada", config=timeout(3))
searchbyscientificname("Quercus douglasii", config=timeout(3))
}
}
\keyword{internal}

