% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchforanymatch}
\alias{searchforanymatch}
\title{Search for any match}
\usage{
searchforanymatch(srchkey = NA, curlopts = list(), curl = getCurlHandle())
}
\description{
Search for any match
}
\examples{
\donttest{
searchforanymatch(srchkey = 202385)
searchforanymatch(srchkey = "dolphin")
}
}
\keyword{internal}

