% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{searchforanymatch}
\alias{searchforanymatch}
\title{Search for any match}
\usage{
searchforanymatch(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for any match
}
\examples{
\dontrun{
searchforanymatch(x=202385, config=timeout(3))
searchforanymatch(x="dolphin", config=timeout(3))
}
}
\keyword{internal}

