% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchforanymatch}
\alias{searchforanymatch}
\title{Search for any match}
\usage{
searchforanymatch(x, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for any match
}
\examples{
\donttest{
searchforanymatch(202385, config=timeout(3))
searchforanymatch("dolphin", config=timeout(3))
}
}
\keyword{internal}

