\name{searchforanymatchpaged}
\alias{searchforanymatchpaged}
\title{match count}
\usage{
  searchforanymatchpaged(srchkey = NA, pagesize = NA,
    pagenum = NA, ascend = NA,
    url = "http://www.itis.gov/ITISWebService/services/ITISService/searchForAnyMatchPaged",
    ..., curl = getCurlHandle())
}
\description{
  match count
}
\keyword{internal}

