% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchforanymatchpaged}
\alias{searchforanymatchpaged}
\title{Search for any matched page}
\usage{
searchforanymatchpaged(x, pagesize = NULL, pagenum = NULL, ascend = NULL,
  ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{pagesize}{An integer containing the page size (numeric)}

\item{pagenum}{An integer containing the page number (numeric)}

\item{ascend}{A boolean containing true for ascending sort order or false
for descending (logical)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Search for any matched page
}
\examples{
\donttest{
searchforanymatchpaged(202385, pagesize=100, pagenum=1, ascend=FALSE, config=timeout(3))
searchforanymatchpaged("Zy", pagesize=100, pagenum=1, ascend=FALSE, config=timeout(3))
}
}
\keyword{internal}

