% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_agg.R
\name{tax_agg}
\alias{tax_agg}
\alias{print.tax_agg}
\title{Aggregate species data to given taxonomic rank}
\usage{
tax_agg(x, rank, db = "ncbi", messages = FALSE, ...)

\method{print}{tax_agg}(x, ...)
}
\arguments{
\item{x}{Community data matrix. Taxa in columns, samples in rows.}

\item{rank}{character; Taxonomic rank to aggregate by.}

\item{db}{character; taxonomic API to use, 'ncbi, 'itis' or both, see
\code{\link[=tax_name]{tax_name()}}. Note that each taxonomic data source has
their own identifiers, so that if you provide the wrong \code{db} value
for the identifier you could get a result, but it will likely be wrong (not
what you were expecting). If using ncbi we recommend getting an API key;
see \link{taxize-authentication}}

\item{messages}{(logical) If FALSE (Default) suppress messages}

\item{...}{Other arguments passed to \code{\link[=get_itis]{get_itis()}} or \code{\link[=get_ncbi]{get_ncbi()}}}
}
\value{
A list of class \code{tax_agg} with the following items:
\itemize{
\item \code{x} Community data matrix with aggregated data.
\item \code{by} A lookup-table showing which taxa were aggregated.
\item \code{n_pre} Number of taxa before aggregation.
\item \code{rank} Rank at which taxa have been aggregated.
}
}
\description{
Aggregate species data to given taxonomic rank
}
\details{
\code{tax_agg} aggregates (sum) taxa to a specific taxonomic level.
If a taxon is not found in the database (ITIS or NCBI) or the supplied taxon
is on higher taxonomic level this taxon is not aggregated.
}
\examples{
\dontrun{
if (requireNamespace("vegan", quietly = TRUE)) {
  # use dune dataset
  data(dune, package='vegan')
  species <- c("Achillea millefolium", "Agrostis stolonifera",
    "Aira praecox", "Alopecurus geniculatus", "Anthoxanthum odoratum",
    "Bellis perennis", "Bromus hordeaceus", "Chenopodium album",
    "Cirsium arvense", "Comarum palustre", "Eleocharis palustris",
    "Elymus repens", "Empetrum nigrum", "Hypochaeris radicata",
    "Juncus articulatus", "Juncus bufonius", "Lolium perenne",
    "Plantago lanceolata", "Poa pratensis", "Poa trivialis",
    "Ranunculus flammula", "Rumex acetosa", "Sagina procumbens",
    "Salix repens", "Scorzoneroides autumnalis", "Trifolium pratense",
    "Trifolium repens", "Vicia lathyroides", "Brachythecium rutabulum",
    "Calliergonella cuspidata")
  colnames(dune) <- species

  # aggregate sample to families
  (agg <- tax_agg(dune, rank = 'family', db = 'ncbi'))

  # extract aggregated community data matrix for further usage
  agg$x
  # check which taxa have been aggregated
  agg$by
}

# A use case where there are different taxonomic levels in the same dataset
spnames <- c('Puma','Ursus americanus','Ursidae')
df <- data.frame(c(1,2,3), c(11,12,13), c(1,4,50))
names(df) <- spnames
out <- tax_agg(x=df, rank = 'family', db='itis')
out$x

# You can input a matrix too
mat <- matrix(c(1,2,3, 11,12,13), nrow = 2, ncol = 3,
 dimnames=list(NULL, c('Puma concolor','Ursus americanus','Ailuropoda melanoleuca')))
tax_agg(mat, rank = 'family', db='itis')
}
}
\seealso{
\link{tax_name}
}
