% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxize-package.R
\name{taxize-defunct}
\alias{taxize-defunct}
\title{Defunct functions in taxize}
\description{
The following functions are now defunct (no longer available):
}
\details{
\itemize{
 \item \code{\link{col_classification}}: See \code{\link{classification}}
 \item \code{\link{eol_hierarchy}}: See \code{\link{classification}}
 \item \code{\link{tp_classification}}: See \code{\link{classification}}
 \item \code{\link{tpl_search}}: Use the \code{Taxonstand} functions
 \code{\link[Taxonstand]{TPL}} or \code{\link[Taxonstand]{TPLck}} directly.
 \item \code{\link{get_seqs}}: This function changed name to \code{\link{ncbi_getbyname}}.
 \item \code{\link{get_genes}}: This function changed name to \code{\link{ncbi_getbyid}}.
 \item \code{\link{get_genes_avail}}: This function changed name to \code{\link{ncbi_search}}.
}
}

