% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize-package.R
\name{taxize-deprecated}
\alias{taxize-deprecated}
\title{Deprecated functions in taxize}
\description{
The following functions are now deprecated:
}
\details{
\itemize{
 \item \code{\link{phylomatic_tree}}: This function is deprecated and will be removed
 in a future version of this package. Use this same function in another package called
 \code{branching} (see \url{https://github.com/ropensci/branching}).
}
}

