% Generated by roxygen2 (4.0.1): do not edit by hand
\name{taxize-deprecated}
\alias{taxize-deprecated}
\title{Deprecated functions in taxize}
\description{
The following functions are now deprecated - see the function \code{classification}
}
\details{
\itemize{
 \item \code{\link{col_classification}}: This function is deprecated. See \code{classification}
 \item \code{\link{eol_hierarchy}}: This function is deprecated. See \code{classification}
 \item \code{\link{tp_classification}}: This function is deprecated. See \code{classification}
}

The following functions have changed names. The old function names are still available to call,
but simply return a message directing the user to the new function name.

\itemize{
 \item \code{\link{get_seqs}}: This function has changed name to \code{ncbi_getbyname}.
 \item \code{\link{get_genes}}: This function has changed name to \code{ncbi_getbyid}.
 \item \code{\link{get_genes_avail}}: This function has changed name to \code{ncbi_search}.
 \item \code{\link{tp_acceptednames}}: This function has changed name to \code{tp_accnames}.
 \item \code{\link{tp_namedistributions}}: This function has changed name to \code{tp_dist}.
 \item \code{\link{tp_namereferences}}: This function has changed name to \code{tp_refs}.
 \item \code{\link{itis_name}}: This function has changed name to \code{tax_name}.
}

In addition, \code{tpl_search} is deprecated - use the Taxonstand functions \code{TPL} or
\code{TPLck} directly.
}

