% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize_options.R
\name{taxize_options}
\alias{taxize_options}
\title{taxize options}
\usage{
taxize_options(taxon_state_messages = NULL, ncbi_sleep = NULL, quiet = FALSE)
}
\arguments{
\item{taxon_state_messages}{(logical) suppress messages? default: \code{NULL}
(same as setting \code{FALSE}). Set to \code{TRUE} to suppress messages, and \code{FALSE}
to not suppress messages}

\item{ncbi_sleep}{(numeric/integer) number of seconds to sleep between
NCBI ENTREZ http requests. applies to the functions: \code{\link[=classification]{classification()}},
\code{\link[=comm2sci]{comm2sci()}}, \code{\link[=genbank2uid]{genbank2uid()}}, \code{\link[=get_uid]{get_uid()}} and \code{\link[=ncbi_children]{ncbi_children()}}.
defaults: 0.334 (without API key) or 0.101 (with API key). minimum value
can not be less than 0.101}

\item{quiet}{(logical) quiet informational output from this function.
default: \code{TRUE}}
}
\description{
taxize options
}
\examples{
\dontrun{
taxize_options()
taxize_options(FALSE)
taxize_options(TRUE)
taxize_options(ncbi_sleep = 0.4)
taxize_options(taxon_state_messages = TRUE, ncbi_sleep = 0.4)
}
}
