% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxize-package.R
\docType{data}
\name{theplantlist}
\alias{theplantlist}
\title{Lookup-table for family, genus, and species names for ThePlantList}
\description{
Lookup-table for family, genus, and species names for ThePlantList
}
\keyword{data}

