% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnrs.R
\name{tnrs}
\alias{tnrs}
\title{Phylotastic Taxonomic Name Resolution Service.}
\usage{
tnrs(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Defunct - service was down far too much to be reliable
}
\keyword{internal}
