% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tnrs_sources.r
\name{tnrs_sources}
\alias{tnrs_sources}
\title{Get sources for the Phylotastic Taxonomic Name Resolution Service.}
\usage{
tnrs_sources(source = NULL)
}
\arguments{
\item{source}{The source to get information on, one of "iPlant_TNRS",
"NCBI", or "MSW3".}
}
\value{
Sources for the TNRS API.
}
\description{
THIS FUNCTION IS DEFUNCT.
}

