% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_accnames.R
\name{tp_accnames}
\alias{tp_accnames}
\title{Return all accepted names for a taxon name with a given id.}
\usage{
tp_accnames(id, key = NULL, ...)
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; loads from .Rprofile.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
List or dataframe.
}
\description{
Return all accepted names for a taxon name with a given id.
}
\examples{
\dontrun{
tp_accnames(id = 25503923)
tp_accnames(id = 25538750)

# No accepted names found
tp_accnames(id = 25509881)
}
}
