% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_dist.R
\name{tp_namedistributions}
\alias{tp_namedistributions}
\title{Return all distribution records for for a taxon name with a given id.}
\usage{
tp_namedistributions(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Function name changed to tp_dist.
}
\keyword{internal}
