% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tp_refs.R
\name{tp_refs}
\alias{tp_refs}
\title{Return all reference records for for a taxon name with a given id.}
\usage{
tp_refs(id, key = NULL, callopts = list())
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; loads from .Rprofile.}

\item{callopts}{Curl options.}
}
\value{
List or dataframe.
}
\description{
Return all reference records for for a taxon name with a given id.
}
\examples{
\dontrun{
tp_refs(id = 25509881)
}
}

