% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_summary.R
\name{tp_summary}
\alias{tp_summary}
\title{Return summary data a taxon name with a given id.}
\usage{
tp_summary(id, key = NULL, ...)
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; See \code{\link{taxize-authentication}} 
for help on authentication}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
A data.frame.
}
\description{
Return summary data a taxon name with a given id.
}
\examples{
\dontrun{
tp_summary(id = 25509881)
tp_summary(id = 2700851)
tp_summary(id = 24900183)
}
}
