% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tp_summary}
\alias{tp_summary}
\title{Return summary data a taxon name with a given id.}
\usage{
tp_summary(id, key = NULL, callopts = list())
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; loads from .Rprofile.}

\item{callopts}{Curl options.}
}
\value{
A data.frame.
}
\description{
Return summary data a taxon name with a given id.
}
\examples{
\donttest{
tp_summary(id = 25509881)
tp_summary(id = 2700851)
tp_summary(id = 24900183)
}
}

