% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_synonyms.R
\name{tp_synonyms}
\alias{tp_synonyms}
\title{Return all synonyms for a taxon name with a given id.}
\usage{
tp_synonyms(id, key = NULL, ...)
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; See \code{\link{taxize-authentication}} 
for help on authentication}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
List or dataframe.
}
\description{
Return all synonyms for a taxon name with a given id.
}
\examples{
\dontrun{
tp_synonyms(id = 25509881)
}
}
