% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tp_synonyms.R
\name{tp_synonyms}
\alias{tp_synonyms}
\title{Return all synonyms for a taxon name with a given id.}
\usage{
tp_synonyms(id, key = NULL, callopts = list())
}
\arguments{
\item{id}{the taxon identifier code}

\item{key}{Your Tropicos API key; loads from .Rprofile.}

\item{callopts}{Further args passed on to httr::GET}
}
\value{
List or dataframe.
}
\description{
Return all synonyms for a taxon name with a given id.
}
\examples{
\dontrun{
tp_synonyms(id = 25509881)
}
}

