% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_families.r
\name{tpl_families}
\alias{tpl_families}
\title{Get The Plant List families.}
\usage{
tpl_families(...)
}
\arguments{
\item{...}{(list) Curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
Returns a \link[tibble:tibble]{tibble::tibble} including the names of all families
indexed by The Plant List, and the major groups into which they fall
(i.e. Angiosperms, Gymnosperms, Bryophytes and Pteridophytes).
}
\description{
Get The Plant List families.
}
\details{
Requires an internet connection in order to connect to
<www.theplantlist.org>.
}
\examples{
\dontrun{
# Get a data.frame of plant families, with the group name 
# (Angiosperms, etc.)
tpl_families()
}
}
\seealso{
\code{\link[=tpl_get]{tpl_get()}}
}
\author{
John Baumgartner (johnbb@student.unimelb.edu.au)
}
