% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ubio_id.R
\name{ubio_id}
\alias{ubio_id}
\title{Search uBio by namebank ID.}
\usage{
ubio_id(namebankID = NULL, keyCode = NULL, ...)
}
\arguments{
\item{namebankID}{(character) uBio namebank ID}

\item{keyCode}{Your uBio API key; loads from .Rprofile. If you don't have
one, obtain one at http://www.ubio.org/index.php?pagename=form.}

\item{...}{Parameters passed on to \code{\link[httr]{GET}}}
}
\value{
A list of four data.frame's, one for the name itself, one for synonyms,
one for vernacular names, and one for citations.
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\examples{
\dontrun{
ubio_id(namebankID = 2483153)
ubio_id(namebankID = 105509)
ubio_id(namebankID = 2843601)
ubio_id(namebankID = 2478181)

# Pass in curl options
library("httr")
ubio_id(namebankID = 2478181, callopts=verbose())
ubio_id(namebankID = 2478181, callopts=timeout(3))
}
}

