% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubio_search.R
\name{ubio_search}
\alias{ubio_search}
\title{This function will return NameBankIDs that match given search terms}
\usage{
ubio_search(...)
}
\arguments{
\item{...}{Parameters, ignored}
}
\description{
THIS FUNCTION IS DEFUNCT.
}
