% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ubio_synonyms.R
\name{ubio_synonyms}
\alias{ubio_synonyms}
\title{Search uBio for taxonomic synonyms by hierarchiesID.}
\usage{
ubio_synonyms(...)
}
\arguments{
\item{...}{Parameters, ignored}
}
\description{
THIS FUNCTION IS DEFUNCT.
}

