% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ubio_synonyms.R
\name{ubio_synonyms}
\alias{ubio_synonyms}
\title{Search uBio for taxonomic synonyms by hierarchiesID.}
\usage{
ubio_synonyms(hierarchiesID = NULL, keyCode = NULL, ...)
}
\arguments{
\item{hierarchiesID}{you must include the hierarchiesID (ClassificationBankID)
to receive the classification synonyms}

\item{keyCode}{Your uBio API key; loads from .Rprofile. If you don't have
one, obtain one at http://www.ubio.org/index.php?pagename=form.}

\item{...}{Parameters passed on to \code{\link[httr]{GET}}}
}
\value{
A data.frame.
}
\description{
THIS FUNCTION IS DEFUNCT.
}
\examples{
\dontrun{
ubio_synonyms(hierarchiesID = 4091702)
ubio_synonyms(hierarchiesID = 2483153)
ubio_synonyms(hierarchiesID = 2465599)
ubio_synonyms(hierarchiesID = 1249021)
ubio_synonyms(hierarchiesID = 4069372)

# Pass in curl options
library("httr")
ubio_synonyms(hierarchiesID = 4091702, config=verbose())
}
}

