% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{function}
\name{worms}
\alias{worms}
\title{Search World Register of Marine Species (WoRMS)}
\description{
WORMS has a SOAP API. We store the machine generated API specification in the package. However,
you can update the spec if you want using \code{}
}
\references{
\url{http://www.marinespecies.org/}
}

