% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_extid}
\alias{worms_extid}
\title{Get external ID from Worms ID}
\usage{
worms_extid(ids = NULL, type = "ncbi", opts = NULL,
  update_iface = FALSE, ...)
}
\arguments{
\item{id}{(numeric) One or more worms AphidID for a taxon.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function
which actually invokes the SOAP method. These options control aspects of the HTTP request,
including debugging information that is displayed on the console,
e.g. .opts = list(verbose = TRUE)}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}

\item{type}{External ID source to get ID for. One of ncbi (default), tsn, bold, eol,
dyntaxa, fishbase, iucn or lsid.}
}
\value{
Character class with ID and attributess for found or not, the source name, and a URL for
the taxon with the external source. No URL is given for lsid.
}
\description{
Get external ID from Worms ID
}
\examples{
\dontrun{
worms_extid(ids=1080, type='ncbi')
worms_extid(ids=278241, type='tsn')
worms_extid(ids=278241, type='iucn')
worms_extid(ids=282981, type='fishbase')
worms_extid(ids=127186, type='bold')
worms_extid(ids=278241, type='eol')
worms_extid(ids=278241, type='lsid')
worms_extid(ids=127186, type='dyntaxa')
}
}

