% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_gen_iface}
\alias{worms_gen_iface}
\title{Generate WORMS function interface.}
\usage{

  worms_gen_iface(wsdl_url = "http://www.marinespecies.org/aphia.php?p=soap&wsdl=1",
  ...)
}
\arguments{
\item{wsdl_url}{URL for the WORMS SOAP WSDL file}

\item{...}{Further args passed on to \code{genSOAPClientInterface}}
}
\value{
Returns invisibly a S4 object holding all functions to interact with WORMS.
}
\description{
Generate WORMS function interface.
}
\examples{
\dontrun{
# You can generate your own interface to WORMS functions
out <- worms_gen_iface()
out

# Then access them directly
out@functions$matchAphiaRecordsByNames

# Or pass to the taxize functions in the iface parameter
head(worms_records(scientific = "Salmo", iface = out))
}
}

