% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_name}
\alias{worms_name}
\title{Get name from a WORMS id}
\usage{
worms_name(id = NULL, opts = NULL, update_iface = FALSE)
}
\arguments{
\item{id}{(numeric) One or more worms AphidID for a taxon.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function
which actually invokes the SOAP method. These options control aspects of the HTTP request,
including debugging information that is displayed on the console,
e.g. .opts = list(verbose = TRUE)}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}

\item{update_iface}{(logical) xx}
}
\description{
Get name from a WORMS id
}
\examples{
\dontrun{
worms_name(id=1080)
}
}

