% Generated by roxygen2 (4.0.1): do not edit by hand
\name{worms_records}
\alias{worms_records}
\title{Get records by ID, scientific name, common name, date, worms id, or external id.}
\usage{
worms_records(scientific = NULL, common = NULL, ids = NULL,
  extids = NULL, like = NULL, type = NULL, fuzzy = FALSE,
  marine_only = 1, offset = NULL, startdate = NULL, enddate = NULL,
  opts = NULL, iface = NULL, ...)
}
\arguments{
\item{ids}{(numeric) One or more WoRMS AphidID's for a taxon.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function
which actually invokes the SOAP method. These options control aspects of the HTTP request,
including debugging information that is displayed on the console,
e.g. .opts = list(verbose = TRUE)}

\item{iface}{Interface to WoRMS SOAP API methods. By default we use a previously created object.
If you want to create a new one, use \code{worms_gen_iface}, assign the output to an object,
then pass it into any \code{worms_*} function. in the \code{iface} parameter.}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}

\item{scientific}{(character) A scientific name.}

\item{common}{(character) A common name.}

\item{extids}{(integer) External identifier.}

\item{like}{(logical) Add a percent sign after the ScientificName (SQL LIKE function). Default=TRUE}

\item{type}{Type of external identifier. Should be one of bold, dyntaxa, eol, fishbase,
iucn, lsid, ncbi, or tsn.}

\item{fuzzy}{(logical) Use fuzzy matching or not. If \code{TRUE}, we use the WoRMS methods
\code{matchAphiaRecordsByNames()} internally for fuzzy/near matching. If \code{FALSE}, we use
the WoRMS method \code{getAphiaRecords()}.}

\item{marine_only}{(logical) Include results from marine taxa only. Default=TRUE.}

\item{offset}{Starting record number, when retrieving next chunk of (50) records. Default=1.}

\item{startdate}{ISO 8601 formatted start date(time). Default=today().
i.e. 2014-08-04T15:57:54+00:00}

\item{enddate}{ISO 8601 formatted start date(time). Default=today().
i.e. 2014-08-04T15:57:54+00:00}
}
\description{
Get records by ID, scientific name, common name, date, worms id, or external id.
}
\details{
Parameter \code{type} should be one of the following values:
\itemize{
 \item bold: Barcode of Life Database (BOLD) TaxID
 \item dyntaxa: Dyntaxa ID
 \item eol: Encyclopedia of Life (EoL) page identifier
 \item fishbase: FishBase species ID
 \item iucn: IUCN Red List Identifier
 \item lsid: Life Science Identifier
 \item ncbi: NCBI Taxonomy ID (Genbank)
 \item tsn: ITIS Taxonomic Serial Number
}
}
\examples{
\dontrun{
worms_records(scientific='Salmo')
worms_records(scientific=c('Salmo','Aphanius'))
worms_records(scientific='Liopsetta glacialis')
worms_records(common='salmon')
worms_records(common=c('salmon','char'))
worms_records(startdate='2014-06-01T00:00:00', enddate='2014-06-02T00:00:00')
worms_records(ids=1080)
worms_records(extids=6830, type='ncbi')

worms_records(scientific="Holothuria edulis")
worms_records(scientific="Holothuria edulis", fuzzy=TRUE)
worms_records(scientific="Holothuria (Halodeima) edulis")

worms_records(scientific='Scotoplanes')

worms_records(scientific='Salmo', offset=51)
}
}

